chore(static site) separate static build from webpack config
