+++ package.json
@@ -25,12 +25,13 @@
     "build-scss": "./scripts/run-node-sass.sh",
     "watch-scss": "./scripts/run-node-sass.sh -w",
     "build-scss-production": "./scripts/run-node-sass.sh -p",
-    "start": "npm run build-scss && npm run watch-scss & npm run watch-js & webpack-dev-server",
+    "start": "node static-build.js && npm run build-scss && npm run watch-scss & npm run watch-js & webpack-dev-server",
     "build-js": "webpack -d",
     "build-js-production": "NODE_ENV=production webpack -p && cat www/bundle.js | gzip > www/bundle.js.gz",
     "watch-js": "webpack -d --watch",
     "build-css": "cat src/base/static/dist/cat-bundle.css | tee www/dist/bundle.css | gzip > www/dist/bundle.css.gz",
     "build-concat": "node scripts/concat.js",
+    "build-dev": "node static-build.js",
     "build": "NODE_ENV=production npm run build-scss-production && npm run build-concat && npm run build-js-production && npm run build-css",
     "prettier": "prettier --write --trailing-comma all \"{src,blah}/**/*.{js,jsx}\""
   },
