+++ src/base/jstemplates/templates.js
@@ -1,0 +1,6 @@
+!function(){var n=Handlebars.template,e=Handlebars.templates=Handlebars.templates||{};e.about=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<h3>Greensboro Participatory Budgeting:</h3>\n<h4>Our Community, Our Decision!</h4>\n<br>\n<p>Greensboro Participatory Budgeting is a democratic process in which community members from all five City Council districts are directly deciding how to spend $500,000 ($100,000 per district).</p>\n\n<p>From April to November 2017, volunteer budget delegates will work to develop these initial ideas into full project proposals. Greensboro residents will able to vote in October 2017 to determine the winning projects.\n</p>\n\n<p>For more information on participatory budgeting in Greensboro, visit <a href="http://www.greensboro-nc.gov/pb" target="_blank">greensboro-nc.gov/pb</a>.</p>\n\n<p>For information on the nationwide movement for PB, visit the website of the <a href="http://www.participatorybudgeting.org/" target="_blank">Participatory Budgeting Project</a>.</p>\n\n<hr>\n<div class="clearfix" style="padding-top:0.5em;">\n  <a href="http://www.participatorybudgeting.org/" target="_blank"><img src="/static/pbp-logo.png" style="float:left;max-width:45%;"></a>\n</div>'},useData:!0}),e["activity-list-item"]=n({1:function(n,e,a,l,t){var o;return n.escapeExpression(n.lambda(null!=(o=null!=e?e.place:e)?o["url-title"]:o,e))},3:function(n,e,a,l,t){var o,i=n.lambda,r=n.escapeExpression;return r(i(null!=(o=null!=e?e.place:e)?o.datasetSlug:o,e))+"/"+r(i(null!=(o=null!=e?e.place:e)?o.id:o,e))},5:function(n,e,a,l,t){var o;return"/response/"+n.escapeExpression(n.lambda(null!=(o=null!=e?e.target:e)?o.id:o,e))},7:function(n,e,a,l,t){var o;return'          <img src="'+n.escapeExpression(n.lambda(null!=(o=null!=(o=null!=e?e.target:e)?o.submitter:o)?o.avatar_url:o,e))+'" class="avatar" />\n'},9:function(n,e,a,l,t){var o;return'          <img src="'+n.escapeExpression((o=null!=(o=a.STATIC_URL||(null!=e?e.STATIC_URL:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"STATIC_URL",hash:{},data:t}):o))+'css/images/user-50.png" class="avatar" />\n'},11:function(n,e,a,l,t){var o;return"          "+n.escapeExpression(n.lambda(null!=(o=null!=(o=null!=e?e.target:e)?o.submitter:o)?o.name:o,e))+"\n"},13:function(n,e,a,l,t){var o;return"          "+n.escapeExpression(n.lambda(null!=(o=null!=e?e.target:e)?o.submitter_name:o,e))+"\n"},15:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{},s=n.escapeExpression;return"          "+s((i=null!=(i=a.action||(null!=e?e.action:e))?i:a.helperMissing,"function"==typeof i?i.call(r,{name:"action",hash:{},data:t}):i))+': <span class="place-type">'+s(n.lambda(null!=(o=null!=e?e.place:e)?o.name:o,e))+"</span>\n\x3c!--           \n"+(null!=(o=a.if.call(r,null!=(o=null!=e?e.place:e)?o.location_text:o,{name:"if",hash:{},fn:n.program(16,t,0),inverse:n.noop,data:t}))?o:"")+" --\x3e\n"},16:function(n,e,a,l,t){var o;return'            at <span class="place-name">'+(null!=(o=n.invokePartial(l["location-string"],null!=(o=null!=e?e.place:e)?o.location_text:o,{name:"location-string",data:t,helpers:a,partials:l,decorators:n.decorators}))?o:"")+"</span>\n"},18:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{};return"          "+n.escapeExpression((i=null!=(i=a.action||(null!=e?e.action:e))?i:a.helperMissing,"function"==typeof i?i.call(r,{name:"action",hash:{},data:t}):i))+":\n"+(null!=(o=a.if.call(r,null!=(o=null!=e?e.place:e)?o.name:o,{name:"if",hash:{},fn:n.program(19,t,0),inverse:n.program(21,t,0),data:t}))?o:"")},19:function(n,e,a,l,t){var o;return'            <span class="place-name">'+n.escapeExpression(n.lambda(null!=(o=null!=e?e.place:e)?o.name:o,e))+"</span>\n"},21:function(n,e,a,l,t){var o,i=n.lambda;return"            a"+(null!=(o=a.blockHelperMissing.call(e,i(null!=(o=null!=e?e.place:e)?o.type_starts_with_vowel:o,e),{name:"place.type_starts_with_vowel",hash:{},fn:n.program(22,t,0),inverse:n.noop,data:t}))?o:"")+" "+n.escapeExpression(i(null!=(o=null!=e?e.place:e)?o.place_type_label:o,e))+"\n"},22:function(n,e,a,l,t){return"n"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{},s=n.escapeExpression;return'      <li class="activity-item clearfix">\n        <a href="/'+(null!=(o=a.if.call(r,null!=(o=null!=e?e.place:e)?o["url-title"]:o,{name:"if",hash:{},fn:n.program(1,t,0),inverse:n.program(3,t,0),data:t}))?o:"")+(null!=(o=a.is.call(r,null!=e?e.target_type:e,"comments",{name:"is",hash:{},fn:n.program(5,t,0),inverse:n.noop,data:t}))?o:"")+'" data-action-type="'+s((i=null!=(i=a.target_type||(null!=e?e.target_type:e))?i:a.helperMissing,"function"==typeof i?i.call(r,{name:"target_type",hash:{},data:t}):i))+'" data-place-id="'+s(n.lambda(null!=(o=null!=e?e.place:e)?o.id:o,e))+'"><strong>\n\n'+(null!=(o=a.if.call(r,null!=(o=null!=(o=null!=e?e.target:e)?o.submitter:o)?o.avatar_url:o,{name:"if",hash:{},fn:n.program(7,t,0),inverse:n.program(9,t,0),data:t}))?o:"")+"\n"+(null!=(o=a.if.call(r,null!=(o=null!=(o=null!=e?e.target:e)?o.submitter:o)?o.name:o,{name:"if",hash:{},fn:n.program(11,t,0),inverse:n.program(13,t,0),data:t}))?o:"")+"\n        </strong>\n\n"+(null!=(o=a.if.call(r,null!=e?e.is_place:e,{name:"if",hash:{},fn:n.program(15,t,0),inverse:n.program(18,t,0),data:t}))?o:"")+"        </a>\n      </li>\n"},usePartial:!0,useData:!0}),e["activity-list"]=n({1:function(n,e,a,l,t){var o;return null!=(o=n.invokePartial(l["activity-list-item"],e,{name:"activity-list-item",data:t,indent:"  ",helpers:a,partials:l,decorators:n.decorators}))?o:""},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s="";return i=null!=(i=a.activities||(null!=e?e.activities:e))?i:a.helperMissing,r={name:"activities",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(null!=e?e:n.nullContext||{},r):i,a.activities||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(s+=o),s},usePartial:!0,useData:!0}),e["add-places"]=n({1:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{};return'\t<div id="main-btns-container" class="pos-top-right">\n'+(null!=(o=a.if.call(i,null!=e?e.add_button_label:e,{name:"if",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:"")+(null!=(o=a.if.call(i,null!=e?e.story_button_label:e,{name:"if",hash:{},fn:n.program(4,t,0),inverse:n.noop,data:t}))?o:"")+"\t</div>\n"},2:function(n,e,a,l,t){var o;return'\t\t  <span id="add-place-btn-container">\n\t\t    <a href="#" id="add-place" class="btn btn-block btn-primary btn-large"><span>'+n.escapeExpression((o=null!=(o=a.add_button_label||(null!=e?e.add_button_label:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"add_button_label",hash:{},data:t}):o))+"</span></a>\n\t\t  </span>\n"},4:function(n,e,a,l,t){var o;return'\t\t  <span id="story-btn-container">\n\t\t    <a href="/page/stories" rel="internal" id="nav-stories" class="btn btn-block btn-primary btn-large"><span>'+n.escapeExpression((o=null!=(o=a.story_button_label||(null!=e?e.story_button_label:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"story_button_label",hash:{},data:t}):o))+"</span></a>\n\t\t  </span>\n"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r;return i=null!=(i=a.adding_supported||(null!=e?e.adding_supported:e))?i:a.helperMissing,r={name:"adding_supported",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(null!=e?e:n.nullContext||{},r):i,a.adding_supported||(o=a.blockHelperMissing.call(e,o,r)),null!=o?o:""},useData:!0}),e["auth-nav"]=n({1:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=n.escapeExpression;return'<nav class="user-menu authed" role="article">\n  <a href="#" id="sign-in-btn"><img class="avatar header-avatar" src="'+r(a.current_user.call(i,"avatar_url",{name:"current_user",hash:{},data:t}))+'"></a>\n  <ul class="menu sign-in-menu">\n    <li><span class="signed-in-as">Signed in as</span> <span class="current-user">'+r(a.current_user.call(i,"name",{name:"current_user",hash:{},data:t}))+'</span></li>\n    <li class=""><a class="logout-btn" href="'+r((o=null!=(o=a.apiRoot||(null!=e?e.apiRoot:e))?o:a.helperMissing,"function"==typeof o?o.call(i,{name:"apiRoot",hash:{},data:t}):o))+'users/logout/">註銷</a>\n    </li>\n  </ul>\n</nav>\n'},3:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'<nav class="user-menu" role="article">\n  <a href="#" id="sign-in-btn">登記</a>\n  <ul class="menu sign-in-menu">\n    <li class="menu-item sign-in-menu-item"><a class="auth-btn twitter-btn" href="'+s((o=null!=(o=a.apiRoot||(null!=e?e.apiRoot:e))?o:r,"function"==typeof o?o.call(i,{name:"apiRoot",hash:{},data:t}):o))+'users/login/twitter/">Twitter</a></li>\n    <li class="menu-item sign-in-menu-item"><a class="auth-btn facebook-btn" href="'+s((o=null!=(o=a.apiRoot||(null!=e?e.apiRoot:e))?o:r,"function"==typeof o?o.call(i,{name:"apiRoot",hash:{},data:t}):o))+'users/login/facebook/">Facebook</a></li>\n  </ul>\n</nav>\n'},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o;return null!=(o=a.is_authenticated.call(null!=e?e:n.nullContext||{},{name:"is_authenticated",hash:{},fn:n.program(1,t,0),inverse:n.program(3,t,0),data:t}))?o:""},useData:!0}),e.background=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<h1>Background</h1>\n\n<p>The Duwamish River is Seattle\'s only river, and is one of the most polluted waterways in the US. After 100 years of urban industrialization, it contains more than 40 toxic chemicals above safe limits for humans and the environment, including PCBs, dioxins, carcinogenic PAHs, and arsenic. Studies show that the low income communities living along the river have higher rates of heart disease, asthma, and other illnesses than other Seattle areas. It is also home to salmon, osprey, sea lions, and other creatures who face unsafe levels of contamination.</p>\n\n<p>The history here is rich. The Duwamish Tribe lived along the shores of this river for over 10,000 years, since the last ice age, until they were removed by incoming settlers who came about a century and a half ago.</p>\n\n<p>In 2001, the US Environmental Protection Agency declared the lower 5 miles of the Duwamish to be a superfund site, which mandates a cleanup paid by those responsible for the pollution. After more than a decade of public deliberation, the EPA came to a final decision in December 2014, and ordered a $342 million cleanup to take place over the next 17 years.</p>\n\n<p>The responsible parties named so far are Boeing, King County, the City of Seattle, and the Port of Seattle, with several more expected to be named over the coming years.</p>\n\n<p>We hope that \'Hey Duwamish\' can be used as a community monitoring system to help ensure that the $342 million is well spent. We want to restore the health of our watershed to its maximum potential and highlight the progress being made through initiatives such as the <a href="http://murray.seattle.gov/murray-announces-funding-for-duwamish-river-community-projects">Duwamish Opportunity Fund</a>.</p>\n\n<p class="btn btn-block btn-large close-btn">Take me to the map!</p>\n'},useData:!0}),e.build=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<h1>How to Build a Rain Garden</h1>\n<p><img class="aligncenter size-full wp-image-646" alt="diagram" src="http://www.12000raingardens.org/wp-content/uploads/2013/01/diagram.jpg" width="676" height="251" /></p>\n<p>If you&#8217;re interested in how to build a rain garden, a step-by-step process can be easily learned by attending a rain garden workshop, studying the <a title="WSU Rain Garden Handbook for Western Washington 2013" href="https://fortress.wa.gov/ecy/publications/publications/1310027.pdf" target="_blank">Washington State University Rain Garden Handbook for Homeowners</a> (18MB PDF download) and viewing the &#8220;How to build a rain garden WSU&#8221; online video (below). You will learn rain garden design and construction basics like:</p>\n<ul>\n<li>How to determine how much impervious surface you have, and how much of that water you want to manage.</li>\n<li>How to decide where on the property you&#8217;d like to build your rain garden.</li>\n<li>How to do a percolation test, to make sure the soils surrounding your rain garden plants can soak up the water.</li>\n<li>How to decide how big of a rain garden to create in the space you have available.</li>\n<li>How to implement rain garden construction next steps.</li>\n<li>How to know what the best rain garden plants are.</li>\n<li>How to maintain a rain garden.</li>\n</ul>\n<p>You don&#8217;t have to be a gardener or an engineer to build a rain garden, but it is important to consider your site carefully before you get started.</p>\n<p><iframe src="http://www.youtube.com/embed/9Kti4HJ45BM" height="315" width="560" allowfullscreen="" frameborder="0"></iframe></p>'},useData:!0}),e.centerpoint=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<span id="centerpoint"><span class="shadow"></span><span class="x"></span><span class="marker"></span></span>'},useData:!0}),e["colorpicker-container"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return"\x3c!-- Template for the Spectrum colorpicker container --\x3e\n\n<div class='toolbar-colorpicker-container'></div>"},useData:!0}),e["colorpicker-controls"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return"\x3c!-- A template for adding additional controls to the Spectrum colorpicker --\x3e\n\n<button type='button' class='sp-choose sp-selected' data-edit-mode='fill'>Fill</button>\n<button type='button' class='sp-choose' data-edit-mode='stroke'>Stroke</button>"},useData:!0}),e.contribute=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<h1>Get involved</h1>\n\n<p>Want to help out? Awesome!</p>\n\n<p>The best way to get involved is to <a href="/new">submit a report</a> on the map, or comment on someone else\'s report.</p>\n\n<p>We could also use help writing code, designing a better interface, and promoting the project. <a href="http://github.com/smartercleanup/platform">Contribute to our repository on GitHub!</a></p>\n\n<p>If you want to see a project like this in your community, let us know!</p>\n\n<a href="http://eepurl.com/6IfA5">Send us a message</a> and we\'ll be in touch ;)\n'},useData:!0}),e["count-meter"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{},s=a.helperMissing,u=n.escapeExpression,c=n.lambda;return'<div class="counter-meter">\n    <div class="col meter-label">\n      '+u((i=null!=(i=a.value_pretty||(null!=e?e.value_pretty:e))?i:s,"function"==typeof i?i.call(r,{name:"value_pretty",hash:{},data:t}):i))+" / "+u((i=null!=(i=a.counter_max_pretty||(null!=e?e.counter_max_pretty:e))?i:s,"function"==typeof i?i.call(r,{name:"counter_max_pretty",hash:{},data:t}):i))+':\n    </div>\n    <div class="col meter-bar">\n      <meter min="'+u(c(null!=(o=null!=e?e.meter_config:e)?o.counter_min:o,e))+'" max="'+u(c(null!=(o=null!=e?e.meter_config:e)?o.counter_max:o,e))+'" low="'+u(c(null!=(o=null!=e?e.meter_config:e)?o.counter_min:o,e))+'" high="'+u(c(null!=(o=null!=e?e.meter_config:e)?o.counter_max:o,e))+'" value="'+u((i=null!=(i=a.value||(null!=e?e.value:e))?i:s,"function"==typeof i?i.call(r,{name:"value",hash:{},data:t}):i))+'" optimum="'+u(c(null!=(o=null!=e?e.meter_config:e)?o.counter_max:o,e))+'" title="rain-gardens"></meter>\n    </div>\n</div>\n'},useData:!0}),e.features=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return"<h1>Features</h1>\n\n<p>For users,</p>\n\n<ul>\n  <li>a simple map interface, to explore submitted locations and add your own\n  <li>site design that works well on small screens\n</ul>\n\n<p>For designers and developers,</p>\n\n<ul>\n  <li>flexible text and layout, making it easy to translate the user interface text or customize the sentences used\n  <li>geospatial database, so you can limit contributions to within boundaries, or asiggn neighborhood names\n</ul>\n\n<p>For project managers during a project,</p>\n\n<ul>\n  <li>simple admin interface\n  <li>blogging tools to create pages or blog posts\n  <li>map and location configuration\n  <li>export to GIS\n  <li>simple status dashboard\n</ul>\n"},useData:!0}),e["filter-menu"]=n({1:function(n,e,a,l,t){var o=n.lambda,i=n.escapeExpression;return'\t<label data-locationtype="'+i(o(null!=e?e.locationType:e,e))+'" \n\t\t\t\t class="filter-menu-item" \n\t\t\t\t for="filter-menu-select-'+i(o(null!=e?e.locationType:e,e))+'">\n\t\t<img src='+i(o(null!=e?e.iconUrl:e,e))+' />\n\t\t<input id="filter-menu-select-'+i(o(null!=e?e.locationType:e,e))+'" \n\t\t\t\t\t type="checkbox" \n\t\t\t\t\t name="filter-menu-select-'+i(o(null!=e?e.locationType:e,e))+'">\n\t\t<span class="filter-menu-item-label">\n\t\t\t'+i(o(null!=e?e.label:e,e))+'\n\t\t</span>\n\t\t<span class="filter-menu-check-icon"></span>\n\t</label>\n'},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o;return'\x3c!-- A template for displaying a menu of location_type filter options --\x3e\n\n<div id="filter-menu-select-reset"\n\t\t data-locationtype="reset" \n\t\t class="filter-menu-item">\n\t<img src="/static/css/images/markers/reset-filter-icon.png" />\n\t<span class="filter-menu-item-label">Reset filters</span>\n</div>\n\n<hr />\n\n'+(null!=(o=a.each.call(null!=e?e:n.nullContext||{},null!=e?e.activeFilters:e,{name:"each",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t}))?o:"")},useData:!0}),e["geocode-address"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o;return'<p class="error is-hidden"></p>\n<form class="geocode-address-form">\n  <label class="geocode-address-label" for="geocode-address-field">搜索</label>\n  <div class="geocode-spinner is-hidden"></div>\n  <input type="text" id="geocode-address-field" class="geocode-address-field" placeholder="'+n.escapeExpression((o=null!=(o=a.geocode_field_label||(null!=e?e.geocode_field_label:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"geocode_field_label",hash:{},data:t}):o))+'">\n</form>\n'},useData:!0}),e["geometry-style-panel"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'\n<div id="geometry-style-control" class="sidebar-tabs">\n  <ul role="tablist" class="sidebar collapsed sidebar-right">\n    <li class="leaflet-control"><i class="fa fa-eyedropper"></i></li>\n    <li class="leaflet-control"><i class="fa fa-eyedropper"></i></li>\n  </ul>\n  <div id="geometry-style-colorpicker"></div>\n</div>'},useData:!0}),e.getinvolved=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<h1>積極參與此活動</h1>\n\n<p>已經提交了看法嗎？有很多方式可以參與預算編制活動：</p>\n\n<h3>志願者</h3>\n\n<p>有很多方式可以參與奧克蘭的PB活動！您可以幫助把資訊傳遞給地區居民，促進小組會議舉行，為PB活動提供支援等。登記<a href="http://www.123contactform.com/form-2317799/PB-Oakland-Volunteer-Sign-Up" target="_blank">這裡</a>。</p>\n\n<h3>聯絡我們</h3>\n\n<p>有問題？請隨時聯絡我們。</p>\n\n<iframe src="http://www.123contactform.com/form-1169694/PB-Oakland-Contact-Us" height="400" width=100% </iframe>'},useData:!0}),e["gis-legend-content"]=n({1:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{},s=a.helperMissing,u=n.escapeExpression;return'          <li class="layer-type-li">\n            <div class="layer-type-title">\n              <input id="map-'+u((i=null!=(i=a.id||(null!=e?e.id:e))?i:s,"function"==typeof i?i.call(r,{name:"id",hash:{},data:t}):i))+'" data-layerid="'+u((i=null!=(i=a.id||(null!=e?e.id:e))?i:s,"function"==typeof i?i.call(r,{name:"id",hash:{},data:t}):i))+'"\n'+(null!=(o=a.if.call(r,null!=e?e.visibleDefault:e,{name:"if",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:"")+'                     class="map-legend-basemap-radio is-visuallyhidden" type="radio" name="basemap"></input>\n              <label for="map-'+u((i=null!=(i=a.id||(null!=e?e.id:e))?i:s,"function"==typeof i?i.call(r,{name:"id",hash:{},data:t}):i))+'">'+u((i=null!=(i=a.title||(null!=e?e.title:e))?i:s,"function"==typeof i?i.call(r,{name:"title",hash:{},data:t}):i))+"</label>\n"+(null!=(o=a.if.call(r,null!=e?e.info:e,{name:"if",hash:{},fn:n.program(4,t,0),inverse:n.program(6,t,0),data:t}))?o:"")+'              <span class="status-icon"></span>\n            </div>\n          </li>\n'},2:function(n,e,a,l,t){return'                     checked="checked"\n'},4:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'                <span class="info-icon" \n                      data-info-content="'+s((o=null!=(o=a.info||(null!=e?e.info:e))?o:r,"function"==typeof o?o.call(i,{name:"info",hash:{},data:t}):o))+'"\n                      data-info-title="'+s((o=null!=(o=a.title||(null!=e?e.title:e))?o:r,"function"==typeof o?o.call(i,{name:"title",hash:{},data:t}):o))+'">\n                </span>\n'},6:function(n,e,a,l,t){return'                <span class="info-icon-placeholder"></span>\n'},8:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=n.escapeExpression,p='    <li class="layer-type-grouping">\n      <div class="layer-type-grouping-title">\n        <input id="'+c((i=null!=(i=a.id||(null!=e?e.id:e))?i:u,"function"==typeof i?i.call(s,{name:"id",hash:{},data:t}):i))+'" class="map-legend-grouping-checkbox is-visuallyhidden"\n               type="checkbox"\n'+(null!=(o=a.if.call(s,null!=e?e.visibleDefault:e,{name:"if",hash:{},fn:n.program(9,t,0),inverse:n.noop,data:t}))?o:"")+'        ></input>\n        <label for="'+c((i=null!=(i=a.id||(null!=e?e.id:e))?i:u,"function"==typeof i?i.call(s,{name:"id",hash:{},data:t}):i))+'">'+c((i=null!=(i=a.title||(null!=e?e.title:e))?i:u,"function"==typeof i?i.call(s,{name:"title",hash:{},data:t}):i))+'</label>\n      </div>\n      <ul class="layer-type-grouping-list unstyled-list">\n';return i=null!=(i=a.layers||(null!=e?e.layers:e))?i:u,r={name:"layers",hash:{},fn:n.program(11,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.layers||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(p+=o),p+"      </ul>\n    </li>\n"},9:function(n,e,a,l,t){return'               checked="checked"\n'},11:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{},s=a.helperMissing,u=n.escapeExpression;return'          <li class="layer-type-li">\n            <div class="layer-type-title">\n              <input id="map-'+u((i=null!=(i=a.id||(null!=e?e.id:e))?i:s,"function"==typeof i?i.call(r,{name:"id",hash:{},data:t}):i))+'" data-layerid="'+u((i=null!=(i=a.id||(null!=e?e.id:e))?i:s,"function"==typeof i?i.call(r,{name:"id",hash:{},data:t}):i))+'"\n'+(null!=(o=a.if.call(r,null!=e?e.visibleDefault:e,{name:"if",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:"")+(null!=(o=a.if.call(r,null!=e?e.constituentLayers:e,{name:"if",hash:{},fn:n.program(12,t,0),inverse:n.noop,data:t}))?o:"")+'                     class="map-legend-checkbox is-visuallyhidden" type="checkbox"></input>\n              <label for="map-'+u((i=null!=(i=a.id||(null!=e?e.id:e))?i:s,"function"==typeof i?i.call(r,{name:"id",hash:{},data:t}):i))+'">'+u((i=null!=(i=a.title||(null!=e?e.title:e))?i:s,"function"==typeof i?i.call(r,{name:"title",hash:{},data:t}):i))+"</label>\n"+(null!=(o=a.if.call(r,null!=e?e.info:e,{name:"if",hash:{},fn:n.program(15,t,0),inverse:n.program(6,t,0),data:t}))?o:"")+'              <span class="status-icon"></span>\n            </div>\n          </li>\n'},12:function(n,e,a,l,t){var o;return'                     data-constituent-layers="'+(null!=(o=a.each.call(null!=e?e:n.nullContext||{},null!=e?e.constituentLayers:e,{name:"each",hash:{},fn:n.program(13,t,0),inverse:n.noop,data:t}))?o:"")+'"\n'},13:function(n,e,a,l,t){return n.escapeExpression(n.lambda(e,e))+" "},15:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'                <span class="info-icon" \n                      data-layerid="'+s((o=null!=(o=a.id||(null!=e?e.id:e))?o:r,"function"==typeof o?o.call(i,{name:"id",hash:{},data:t}):o))+'"\n                      data-info-content="'+s((o=null!=(o=a.info||(null!=e?e.info:e))?o:r,"function"==typeof o?o.call(i,{name:"info",hash:{},data:t}):o))+'"\n                      data-info-title="'+s((o=null!=(o=a.title||(null!=e?e.title:e))?o:r,"function"==typeof o?o.call(i,{name:"title",hash:{},data:t}):o))+'">\n                </span>\n'},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=a.blockHelperMissing,p='\x3c!-- MAP LAYERS --\x3e\n\n<ul class="layer-type-list unstyled-list">\n    <li class="layer-type-grouping">\n      <div class="layer-type-grouping-title">Basemaps</div>\n      <ul class="layer-type-grouping-list unstyled-list">\n';return i=null!=(i=a.basemaps||(null!=e?e.basemaps:e))?i:u,r={name:"basemaps",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.basemaps||(o=c.call(e,o,r)),null!=o&&(p+=o),p+="      </ul>\n    </li>\n",i=null!=(i=a.groupings||(null!=e?e.groupings:e))?i:u,r={name:"groupings",hash:{},fn:n.program(8,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.groupings||(o=c.call(e,o,r)),null!=o&&(p+=o),p+"</ul>\n"},useData:!0}),e.greenscreen=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'    <p>Hello residents of Georgetown and South Park,</p>\n    \n    <p>Just Health Action (JHA), a non-profit Seattle-based organization\n    received a grant from King County Waste Water Treatment Division to build\n    an actionable project in the neighborhood to mitigate air pollution.  This\n    action-oriented project is called Green Solutions to Air Pollution and will\n    leverage the existing resources and work achieved through the Duwamish\n    Community Action for Clean Air project coordinated by Duwamish River\n    Cleanup Coalition (DRCC). We will be partnering with the DRCC and Urban\n    Systems Design’s program “Duwamish Infrastructure Restoration Training”\n    (DIRT Corps) to implement the project.\n     \n<p>After talking to a few residents at community workshops this fall about different types of actionable projects that could reduce air pollution in the neighborhoods, we heard you say that installing a “green screen” was the most exciting option (a detailed description of what a green screen is\nbelow). We now need your feedback about where you would like this green screen to go.  Though we have spoken to some residents already, we want to make sure we hear as many voices as possible before deciding on the final locations (one location per neighborhood). More information about the details of the project are listed below the instructions.</p>\n    \n    <p><strong>To add your green screen location nomination to the map:</strong></p>\n\n<ul>\n<li>Hit the orange button that says \'Add a report to the map\'</li>\n<li>Move the red pin to mark your exact desired location (by dragging and zooming the map)</li>\n<li>Select <em>Green Screen Vote</em> under the dropdown menu where it says \'type of report.\'</li>\n</ul>\n\n    <a href="#" id="add-place" class="btn btn-block btn-primary btn-large"><span>Add a report to the map!</span></a>\n    <img src="/static/css/images/marker-drop_screenshot.jpg" alt="Location pin">\n    <img src="/static/css/images/screenshot_greenscreen_dropdown.jpg" alt="Green Screen Vote - Dropdown Menu">\n  <br>    \n    <p>Feel free to comment on other Green Screen votes that others have chosen\n    to help guide the community process. To comment on other nominations just\n    click on the pin and a comment window will open to the right.</p>\n    \n    For those folks not familiar with the project, here is some more\n    information.</p>\n<img src="/static/css/images/greenscreen_example.png" alt="Green screen example" style="float:left;margin:0 5px 0 0;" />  \n    <h3>What is a green screen?</h3>\n    <p>A green screen is a metal frame structure that will support the growth\n    of native and fast growing vines (no invasive species will be used).  In\n    two to three years the entire structure will be completely covered by the\n    vines, and as the vines grow, so will their ability to absorb air\n    pollution.  It will end up looking like a ten-foot tall and twenty to\n    thirty foot long wall of green growing things.  An example of a green\n    screen on the Goat Hill Parking Garage in Seattle can be found <a href="http://greenscreen.com/docs/Projects%20PDFs/All%20Individual%20Project%20PDFs/Goat%20Hill%20Parking.pdf">here</a>.\n\n    <h3>Why a “green screen”?</h3>\n    <p>A green screen is a great tool to help absorb air pollution.  It is a\n    version of the green wall or living wall concept that requires very little\n    maintenance, so it has the potential to be sustainable for many years to\n    come.  The project is something that can be constructed relatively quickly\n    on a narrow piece of land that cannot accommodate trees. The green screen\n    will look attractive once it’s grown in, as well as reduce noise pollution,\n    absorb water, reduce heat, create local jobs, and it is generally good for\n    your health. Here is a link to more information about the benefits of\n    living walls general:\n    <a href="http://www.greenovergrey.com/green-wall-benefits/overview.php">http://www.greenovergrey.com/green-wall-benefits/overview.php</a></p>\n     \n    <h3>What are we asking you to do?</h3>\n    <p>We would like your input on where you think this green screen should be\n    placed.  To add your voice, go to HeyDuwamish.org and place a pin on\n    the map where you would like the green screen to go (see above for more\n    detailed instructions).  You can also add a comment along with your map pin\n    on why you picked this spot, and/or comment on other people’s pins. \n     Please comment by the end of December. </p>\n     \n    <h3>What is the timeline and how will the project get built?</h3>\n    <p>Two projects will be built, one in Georgetown, and one in South Park. \n    During the first week of January, we will take all of the pins and\n    suggestions into account, and decide on a location for the green screen. \n    We will then negotiate with the landowner or the city as the case may be to\n    site the project on their property. We are aiming to begin construction in\n    the winter or spring of 2016.  The project will be built by the Duwamish\n    Valley Youth Corps, and the Duwamish Infrastructure Restoration Training\n    Corps (DIRT Corps)</p>\n     \n    <p>If you have any additional questions, or would like more information,\n    you can contact us at:</p>\n    <p>Linn Gould - gouldjha@gmail.com</p>\n    <p>Andrew Schiffer - bricktree66@gmail.com</p>\n\n<img src="/static/css/images/kclogo.gif" alt="King County">\n<img src="/static/css/images/jha-logo.jpg" alt="Just Health Action">\n<img src="/static/css/images/dirtcorps-logo.jpg" alt="Dirt Corps">\n<img src="/static/css/images/drcc-logo_small.jpg" alt="Just Health Action">'},useData:!0}),e.guidelines=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'\n<iframe src="https://player.vimeo.com/video/84311420?color=ff9933&byline=0" width="640" height="360" frameborder="0" webkitallowfullscreen mozallowfullscreen allowfullscreen></iframe>\n\n\n<h2>屋崙參與式預算制定究竟如何運作？</h2>\n<p>\n參與式預算制定(PB) 是一項富於革新精神的民主進程，它授權屋崙市議會第一和第二選區的居民就聯邦社區發展綜合補助金(CDBG)應如何使用訂立優先順序，用以改善他們區内低至中等收入的社區。 第二區2017-19年的社區發展綜合補助金（ CDBG）每年撥款$278,685 ，為期兩年，亦即總金額$557,370。那些獲選票最多的項目，算至撥款金額完盡為止，將成為本區社區發展綜合補助金的優先項目。屋崙市政府將發佈一項「提交建議書邀請」(RFP)，然後，市及非牟利機構可以申請撥款來貫徹那些優先項目。社區發展區委員會的委員將審查所有回應「提交建議書邀請」的申請書，並向市府建議哪些申請應獲撥款。\n</p>\n'},useData:!0}),e.instructions=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return"<h1>Detailed instructions</h1>"},useData:!0}),e.involved=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){
+return'<h1>Get More Involved</h1>\n\n<p>Already submitted an idea? There are many more ways to get involved in participatory budgeting:</p>\n\n<h5>Attend a Neighborhood Assembly</h5>\n<p> Each district will have a neighborhood assembly and other, informal idea collection events, to encourage widespread participation. <a href="http://www.greensboro-nc.gov/index.aspx?page=5084">Click here</a> for upcoming assembly dates, times, and locations.</p>\n\n<h5>Connect with Us</h5>\n<p>Find us on <a href="https://www.facebook.com/PBGreensboro/" target="_blank"> Facebook</a> and follow us on <a href="https://twitter.com/GreensboroPB" target="_blank"> Twitter</a>!</p>\n\n<h5>Contact Us</h5>\n<p><strong>Have a question?</strong><br>\nPlease contact Valerie Warren, Community Outreach Coordinator for Participatory Budgeting Greensboro, by email at <a href="mailto:valerie@participatorybudgeting.org">valerie@participatorybudgeting.org</a> or phone at 336‑373‑7750.\n</p>'},useData:!0}),e["layer-info-window"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{},s=a.helperMissing;return'\x3c!-- A template for rendering popup layer info windows --\x3e\n<div class="caret-left"></div>\n<div class="layer-info-window-content">\n\t<div class="info-window-header">\n\t\t<div class="info-window-close-btn">&#10005;</div>\n\t\t<div class="info-window-title">'+n.escapeExpression((i=null!=(i=a.title||(null!=e?e.title:e))?i:s,"function"==typeof i?i.call(r,{name:"title",hash:{},data:t}):i))+'</div>\n\t\t<hr />\n\t</div>\n\t<div class="info-window-body-container">\n\t\t<div class="info-window-body">'+(null!=(i=null!=(i=a.body||(null!=e?e.body:e))?i:s,o="function"==typeof i?i.call(r,{name:"body",hash:{},data:t}):i)?o:"")+"</div>\n\t</div>\n</div>\n"},useData:!0}),e.legend=n({1:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=n.escapeExpression,m='    <li class="shareabouts-layer-li legend-layer-li" >\n        <img class="shareabouts-layer-title" location="'+p((i=null!=(i=a.title||(null!=e?e.title:e))?i:u,typeof i===c?i.call(s,{name:"title",hash:{},data:t}):i))+'" layer-title="'+p((i=null!=(i=a.title||(null!=e?e.title:e))?i:u,typeof i===c?i.call(s,{name:"title",hash:{},data:t}):i))+'" src="'+p((i=null!=(i=a.image||(null!=e?e.image:e))?i:u,typeof i===c?i.call(s,{name:"image",hash:{},data:t}):i))+'">\n        <span>'+p((i=null!=(i=a.title||(null!=e?e.title:e))?i:u,typeof i===c?i.call(s,{name:"title",hash:{},data:t}):i))+'</span>\n        <ul class="nested-legend-list">\n';return i=null!=(i=a.items||(null!=e?e.items:e))?i:u,r={name:"items",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a.items||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(m+=o),m+"        </ul>\n\n    </li>\n"},2:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'            <li class="legend-layer-li">\n              <a class="shareabouts-layer-li" href="'+s((o=null!=(o=a.url||(null!=e?e.url:e))?o:r,"function"==typeof o?o.call(i,{name:"url",hash:{},data:t}):o))+'">'+s((o=null!=(o=a.title||(null!=e?e.title:e))?o:r,"function"==typeof o?o.call(i,{name:"title",hash:{},data:t}):o))+"</a>\n            </li>\n"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s='\x3c!-- SHAREABOUTS LAYERS --\x3e\n<ul class="master-layer-list">\n';return i=null!=(i=a.items||(null!=e?e.items:e))?i:a.helperMissing,r={name:"items",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(null!=e?e:n.nullContext||{},r):i,a.items||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(s+=o),s+"\n</ul>\n"},useData:!0}),e["location-string"]=n({1:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{};return(null!=(o=a.is.call(i,null!=e?e.geocodeQuality:e,"POINT",{name:"is",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:"")+"\n"+(null!=(o=a.is.call(i,null!=e?e.geocodeQuality:e,"ADDRESS",{name:"is",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:"")+"\n"+(null!=(o=a.is.call(i,null!=e?e.geocodeQuality:e,"ZIP",{name:"is",hash:{},fn:n.program(4,t,0),inverse:n.noop,data:t}))?o:"")+"\n"+(null!=(o=a.is.call(i,null!=e?e.geocodeQuality:e,"CITY",{name:"is",hash:{},fn:n.program(6,t,0),inverse:n.noop,data:t}))?o:"")+"\n"+(null!=(o=a.is.call(i,null!=e?e.geocodeQuality:e,"STREET",{name:"is",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:"")},2:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return"    "+s((o=null!=(o=a.street||(null!=e?e.street:e))?o:r,"function"==typeof o?o.call(i,{name:"street",hash:{},data:t}):o))+", "+s((o=null!=(o=a.adminArea5||(null!=e?e.adminArea5:e))?o:r,"function"==typeof o?o.call(i,{name:"adminArea5",hash:{},data:t}):o))+" "+s((o=null!=(o=a.adminArea3||(null!=e?e.adminArea3:e))?o:r,"function"==typeof o?o.call(i,{name:"adminArea3",hash:{},data:t}):o))+"\n"},4:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return"    "+s((o=null!=(o=a.adminArea5||(null!=e?e.adminArea5:e))?o:r,"function"==typeof o?o.call(i,{name:"adminArea5",hash:{},data:t}):o))+", "+s((o=null!=(o=a.adminArea3||(null!=e?e.adminArea3:e))?o:r,"function"==typeof o?o.call(i,{name:"adminArea3",hash:{},data:t}):o))+" "+s((o=null!=(o=a.postalCode||(null!=e?e.postalCode:e))?o:r,"function"==typeof o?o.call(i,{name:"postalCode",hash:{},data:t}):o))+"\n"},6:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return"    "+s((o=null!=(o=a.adminArea5||(null!=e?e.adminArea5:e))?o:r,"function"==typeof o?o.call(i,{name:"adminArea5",hash:{},data:t}):o))+", "+s((o=null!=(o=a.adminArea3||(null!=e?e.adminArea3:e))?o:r,"function"==typeof o?o.call(i,{name:"adminArea3",hash:{},data:t}):o))+"\n"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},e,{name:"if",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t}))?o:""},useData:!0}),e.overview=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<h1>Welcome</h1>\n\n\x3c!-- <p>We are here to educate and inspire the environmental stewardship of Longfellow Creek in West Seattle, starting with Roxhill Bog, all the way down to Elliot Bay!</p>\n<a href="/longfellow" rel="internal">\n<img src="https://5937delridge.files.wordpress.com/2011/10/beaver-pond1.jpg">\n<p class="btn btn-block btn-large close-btn">Start Mapping!</p>\n</a>\n<br>\n<br>\nThis project made possible by our generous <a href="/page/sponsors">sponsors</a>. --\x3e\n\n<p>We are here to educate and inspire action for our Central Puget Sound watersheds</p>\n\x3c!-- \n<figure class="snip1579">\n\t<img src="/static/css/images/overview-image-1.jpg"/>\n\t<figcaption>\n\t  <h3>Lake Sammamish Watershed</h3>\n\t  <blockquote>\n\t    <p></p>\n\t  </blockquote>\n\t</figcaption><a href="/sammamish" rel="internal"></a>\n</figure>\n<figure class="snip1579">\n\t<img src="/static/css/images/overview-image-2.jpg"/>\n  <figcaption>\n    <h3>Cedar River / Lake Washington Watershed</h3>\n    <blockquote>\n      <p></p>\n    </blockquote>\n  </figcaption><a href="/cedar" rel="internal"></a>\n</figure>\n<figure class="snip1579">\n\t<img src="/static/css/images/overview-image-3.jpg" alt="tools"/>\n  <figcaption>\n    <h3>Green / Duwamish Watershed</h3>\n    <blockquote>\n      <p></p>\n    </blockquote>\n  </figcaption><a href="/green-duwamish" rel="internal"></a>\n</figure>\n<figure class="snip1579">\n  <img src="/static/css/images/overview-image-4.jpg" alt="tools"/>\n  <figcaption>\n    <h3>Puyallup / White Watershed</h3>\n    <blockquote>\n      <p></p>\n    </blockquote>\n  </figcaption><a href="/puyallup-white" rel="internal"></a>\n</figure>\n<br><br>\n --\x3e\n\n<figure class="flow-layout flow-sq flow-sq-left">\n  <a href="/cedar" rel="internal">\n    <img src="/static/css/images/watershed-cedar-sq.jpg"/>\n    <figcaption class="flow-layout-title flow-layout-title-bottom">\n      <h3>Cedar River / Lake Washington Watershed</h3>\n    </figcaption>\n  </a>\n</figure>\n\n<figure class="flow-layout flow-vert flow-vert-right">\n  <a href="/green-duwamish" rel="internal">\n    <img src="/static/css/images/watershed-green-duwamish-vert-b.jpg"/>\n    <figcaption class="flow-layout-title flow-layout-title-bottom">\n      <h3>Green / Duwamish Watershed</h3>\n    </figcaption>\n  </a>\n</figure>\n\n<figure class="flow-layout flow-sq flow-sq-left">\n  <a href="/sammamish" rel="internal">\n    <img src="/static/css/images/watershed-sammamish-alt-sq.jpg"/>\n    <figcaption class="flow-layout-title flow-layout-title-bottom">\n      <h3>Lake Sammamish Watershed</h3>\n    </figcaption>\n  </a>\n</figure>\n\n<div class="clearfix"></div>\n\n<figure class="flow-layout">\n  <a href="/puyallup-white" rel="internal">\n    <img src="/static/css/images/watershed-puyallup-alt-horiz.png"/>\n    <figcaption class="flow-layout-title flow-layout-title-bottom">\n      <h3>Puyallup / White Watershed</h3>\n    </figcaption>\n  </a>\n</figure>\n\n<br />\n\nThis project made possible by our generous <a href="/page/sponsors" rel="internal">sponsors</a>.'},useData:!0}),e["page-nav-item"]=n({1:function(n,e,a,l,t){var o;return" internal-menu-item menu-item-"+n.escapeExpression((o=null!=(o=a.slug||(null!=e?e.slug:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"slug",hash:{},data:t}):o))},3:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'    <a class="btn btn-block btn-secondary" href="'+s((o=null!=(o=a.url||(null!=e?e.url:e))?o:r,"function"==typeof o?o.call(i,{name:"url",hash:{},data:t}):o))+'">'+s((o=null!=(o=a.title||(null!=e?e.title:e))?o:r,"function"==typeof o?o.call(i,{name:"title",hash:{},data:t}):o))+"</a>\n"},5:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=a.blockHelperMissing,p="";return i=null!=(i=a.slug||(null!=e?e.slug:e))?i:u,r={name:"slug",hash:{},fn:n.program(6,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.slug||(o=c.call(e,o,r)),null!=o&&(p+=o),i=null!=(i=a.slug||(null!=e?e.slug:e))?i:u,r={name:"slug",hash:{},fn:n.noop,inverse:n.program(8,t,0),data:t},o="function"==typeof i?i.call(s,r):i,a.slug||(o=c.call(e,o,r)),null!=o&&(p+=o),p},6:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'      <a class="btn btn-block btn-secondary" href="/page/'+s((o=null!=(o=a.slug||(null!=e?e.slug:e))?o:r,"function"==typeof o?o.call(i,{name:"slug",hash:{},data:t}):o))+'">'+s((o=null!=(o=a.title||(null!=e?e.title:e))?o:r,"function"==typeof o?o.call(i,{name:"title",hash:{},data:t}):o))+"</a>\n"},8:function(n,e,a,l,t){var o;return"      "+n.escapeExpression((o=null!=(o=a.title||(null!=e?e.title:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"title",hash:{},data:t}):o))+"\n"},10:function(n,e,a,l,t){var o;return null!=(o=n.invokePartial(l["page-nav-item"],e,{name:"page-nav-item",data:t,indent:"    ",helpers:a,partials:l,decorators:n.decorators}))?o:""},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=a.blockHelperMissing,p='<li class="menu-item';return i=null!=(i=a.slug||(null!=e?e.slug:e))?i:u,r={name:"slug",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.slug||(o=c.call(e,o,r)),null!=o&&(p+=o),p+='">\n\n',i=null!=(i=a.external||(null!=e?e.external:e))?i:u,r={name:"external",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.external||(o=c.call(e,o,r)),null!=o&&(p+=o),p+="\n",i=null!=(i=a.external||(null!=e?e.external:e))?i:u,r={name:"external",hash:{},fn:n.noop,inverse:n.program(5,t,0),data:t},o="function"==typeof i?i.call(s,r):i,a.external||(o=c.call(e,o,r)),null!=o&&(p+=o),p+="\n  <ul>\n",i=null!=(i=a.pages||(null!=e?e.pages:e))?i:u,r={name:"pages",hash:{},fn:n.program(10,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.pages||(o=c.call(e,o,r)),null!=o&&(p+=o),p+"  </ul>\n\n</li>\n"},usePartial:!0,useData:!0}),e["pages-nav-item"]=n({1:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{};return(null!=(o=a.unless.call(r,null!=e?e.pages:e,{name:"unless",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:"")+" menu-item-"+n.escapeExpression((i=null!=(i=a.slug||(null!=e?e.slug:e))?i:a.helperMissing,"function"==typeof i?i.call(r,{name:"slug",hash:{},data:t}):i))},2:function(n,e,a,l,t){return" internal-menu-item"},4:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c='\n    <a class="btn btn-block btn-secondary">'+n.escapeExpression((i=null!=(i=a.title||(null!=e?e.title:e))?i:u,"function"==typeof i?i.call(s,{name:"title",hash:{},data:t}):i))+"</a>\n    <ul>\n";return i=null!=(i=a.pages||(null!=e?e.pages:e))?i:u,r={name:"pages",hash:{},fn:n.program(5,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.pages||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(c+=o),c+"    </ul>\n\n"},5:function(n,e,a,l,t){var o;return null!=(o=n.invokePartial(l["pages-nav-item"],e,{name:"pages-nav-item",data:t,indent:"      ",helpers:a,partials:l,decorators:n.decorators}))?o:""},7:function(n,e,a,l,t){var o;return"\n"+(null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.external:e,{name:"if",hash:{},fn:n.program(8,t,0),inverse:n.program(10,t,0),data:t}))?o:"")+"\n"},8:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'      <a class="btn btn-block btn-secondary" href="'+s((o=null!=(o=a.url||(null!=e?e.url:e))?o:r,"function"==typeof o?o.call(i,{name:"url",hash:{},data:t}):o))+'">'+s((o=null!=(o=a.title||(null!=e?e.title:e))?o:r,"function"==typeof o?o.call(i,{name:"title",hash:{},data:t}):o))+"</a>\n"},10:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.slug:e,{name:"if",hash:{},fn:n.program(11,t,0),inverse:n.program(13,t,0),data:t}))?o:""},11:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'        <a class="btn btn-block btn-secondary" href="/page/'+s((o=null!=(o=a.slug||(null!=e?e.slug:e))?o:r,"function"==typeof o?o.call(i,{name:"slug",hash:{},data:t}):o))+'">'+s((o=null!=(o=a.title||(null!=e?e.title:e))?o:r,"function"==typeof o?o.call(i,{name:"title",hash:{},data:t}):o))+"</a>\n"},13:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'        <a class="btn btn-block btn-secondary" href="/'+s((o=null!=(o=a.name||(null!=e?e.name:e))?o:r,"function"==typeof o?o.call(i,{name:"name",hash:{},data:t}):o))+'" rel="internal">'+s((o=null!=(o=a.title||(null!=e?e.title:e))?o:r,"function"==typeof o?o.call(i,{name:"title",hash:{},data:t}):o))+"</a>\n"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{};return'<li class="menu-item'+(null!=(o=a.if.call(i,null!=e?e.slug:e,{name:"if",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t}))?o:"")+'">\n'+(null!=(o=a.if.call(i,null!=e?e.pages:e,{name:"if",hash:{},fn:n.program(4,t,0),inverse:n.program(7,t,0),data:t}))?o:"")+"\n</li>\n"},usePartial:!0,useData:!0}),e["pages-nav"]=n({1:function(n,e,a,l,t){return'<a href="#" id="nav-btn">&equiv;</a>\n'},3:function(n,e,a,l,t){var o;return"\n"+(null!=(o=n.invokePartial(l["pages-nav-item"],e,{name:"pages-nav-item",data:t,indent:"      ",helpers:a,partials:l,decorators:n.decorators}))?o:"")+"\n"},5:function(n,e,a,l,t){var o;return"\n"+(null!=(o=n.invokePartial(l["pages-nav-item"],e,{name:"pages-nav-item",data:t,indent:"        ",helpers:a,partials:l,decorators:n.decorators}))?o:"")+"\n"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=a.blockHelperMissing,m=n.escapeExpression,h="";return i=null!=(i=a.has_pages||(null!=e?e.has_pages:e))?i:u,r={name:"has_pages",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a.has_pages||(o=p.call(e,o,r)),null!=o&&(h+=o),h+='\n\x3c!-- Begin flavor-specific code --\x3e\n\x3c!-- <nav class="add-place-nav">\n  <span id="add-place-btn-container">\n    <a href="#" id="add-place" class="btn btn-block btn-primary">\n      <span>'+m((i=null!=(i=a.add_button_label||(null!=e?e.add_button_label:e))?i:u,typeof i===c?i.call(s,{name:"add_button_label",hash:{},data:t}):i))+'</span>\n    </a>\n  </span>\n</nav> --\x3e\n\x3c!-- End flavor-specific code --\x3e\n\n<nav class="list-toggle-nav">\n  \x3c!-- Begin flavor-specific code --\x3e\n  <a class="list-toggle-btn btn btn-block btn-secondary">\n  \x3c!-- End flavor-specific code --\x3e\n    <span class="show-the-map is-visuallyhidden">'+m((i=null!=(i=a.show_map_button_label||(null!=e?e.show_map_button_label:e))?i:u,typeof i===c?i.call(s,{name:"show_map_button_label",hash:{},data:t}):i))+'</span>\n    <span class="is-screen-reader-text">/</span>\n    <span class="show-the-list">'+m((i=null!=(i=a.show_list_button_label||(null!=e?e.show_list_button_label:e))?i:u,typeof i===c?i.call(s,{name:"show_list_button_label",hash:{},data:t}):i))+'</span>\n  </a>\n</nav>\n\n<nav class="access" role="article">\n  <ul class="menu">\n',i=null!=(i=a.pages||(null!=e?e.pages:e))?i:u,r={name:"pages",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a.pages||(o=p.call(e,o,r)),null!=o&&(h+=o),h+='  </ul>\n</nav>\n\n\x3c!-- Begin flavor-specific code --\x3e\n<div id="pages-nav-right-container">\n  <nav class="access" role="article">\n    <ul class="menu">\n',i=null!=(i=a.pagesRight||(null!=e?e.pagesRight:e))?i:u,r={name:"pagesRight",hash:{},fn:n.program(5,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a.pagesRight||(o=p.call(e,o,r)),null!=o&&(h+=o),h+"    </ul>\n  </nav>\n</div>\n\x3c!-- End flavor-specific code --\x3e"},usePartial:!0,useData:!0}),e["place-detail-edit-bar"]=n({1:function(n,e,a,l,t){var o;return'  <div class="place-detail-edit-bar">\n    <div id="toggle-editor-btn" \n         class="btn btn-small editor-btn">\n      <i class="fa fa-pencil" \n         aria-hidden="true"></i> Edit this post\n    </div>\n\n'+(null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.isEditingToggled:e,{name:"if",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:"")+"  </div>\n"},2:function(n,e,a,l,t){var o;return'      <div id="hide-place-model-btn" \n           class="btn btn-small editor-btn btn-delete">\n        <i class="fa fa-times" \n           aria-hidden="true"></i> Remove\n      </div>\n      <div id="update-place-model-btn" \n           class="btn btn-small editor-btn btn-update'+(null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.isModified:e,{name:"if",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t}))?o:"")+'">\n        <i class="fa fa-floppy-o" \n           aria-hidden="true"></i> Save\n      </div>\n'},3:function(n,e,a,l,t){return" isModified"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o;return"\x3c!-- Template for the row of editor buttons visible at the top of place\n    detail views. --\x3e\n\n"+(null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.isEditable:e,{name:"if",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t}))?o:"")},useData:!0}),e["place-detail-field-types"]=n({1:function(n,e,a,l,t,o,i){var r=n.escapeExpression;return'  <p class="place-value place-value-'+r(n.lambda(null!=i[1]?i[1].name:i[1],e))+'">'+r(a.nlToBr.call(null!=e?e:n.nullContext||{},null!=e?e.content:e,{name:"nlToBr",hash:{},data:t}))+"</p>\n"},3:function(n,e,a,l,t,o,i){var r;return'  <p class="place-value place-value-'+n.escapeExpression(n.lambda(null!=i[1]?i[1].name:i[1],e))+'">'+(null!=(r=a.nlToBr.call(null!=e?e:n.nullContext||{},null!=e?e.content:e,{name:"nlToBr",hash:{},data:t}))?r:"")+"</p>\n"},5:function(n,e,a,l,t,o,i){var r,s;return'  <span class="place-value place-value-'+n.escapeExpression(n.lambda(null!=i[1]?i[1].name:i[1],e))+'">'+(null!=(s=null!=(s=a.content||(null!=e?e.content:e))?s:a.helperMissing,r="function"==typeof s?s.call(null!=e?e:n.nullContext||{},{name:"content",hash:{},data:t}):s)?r:"")+"</span>\n"},7:function(n,e,a,l,t){var o;return"  <ul>\n"+(null!=(o=a.each.call(null!=e?e:n.nullContext||{},null!=e?e.content:e,{name:"each",hash:{},fn:n.program(8,t,0),inverse:n.noop,data:t}))?o:"")+"  </ul>\n"},8:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.selected:e,{name:"if",hash:{},fn:n.program(9,t,0),inverse:n.noop,data:t}))?o:""},9:function(n,e,a,l,t){var o;return'        <li class="multi-select-response">'+n.escapeExpression((o=null!=(o=a.label||(null!=e?e.label:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"label",hash:{},data:t}):o))+" </li>\n"},11:function(n,e,a,l,t){var o;return"  <p>\n"+(null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=(o=null!=e?e.content:e)?o.selected:o,{name:"if",hash:{},fn:n.program(12,t,0),inverse:n.program(14,t,0),data:t}))?o:"")+"  </p>\n"},12:function(n,e,a,l,t){var o;return"      "+n.escapeExpression(n.lambda(null!=(o=null!=e?e.content:e)?o.selectedLabel:o,e))+"\n"},14:function(n,e,a,l,t){var o;return"      "+n.escapeExpression(n.lambda(null!=(o=null!=e?e.content:e)?o.unselectedLabel:o,e))+"\n"},16:function(n,e,a,l,t){var o;return null!=(o=a.each.call(null!=e?e:n.nullContext||{},null!=e?e.content:e,{name:"each",hash:{},fn:n.program(17,t,0),inverse:n.noop,data:t}))?o:""},17:function(n,e,a,l,t){var o;return"    <p>\n"+(null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.selected:e,{name:"if",hash:{},fn:n.program(18,t,0),inverse:n.noop,data:t}))?o:"")+"    </p>\n"},18:function(n,e,a,l,t){return"        "+n.escapeExpression(a.nlToBr.call(null!=e?e:n.nullContext||{},null!=e?e.label:e,{name:"nlToBr",hash:{},data:t}))+"\n"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t,o,i){var r,s=null!=e?e:n.nullContext||{};return"\x3c!-- Template for all fields displayed by the place detail view. --\x3e\n\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"datetime",{name:"is",hash:{},fn:n.program(1,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"text",{name:"is",hash:{},fn:n.program(1,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"textarea",{name:"is",hash:{},fn:n.program(3,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"richTextarea",{name:"is",hash:{},fn:n.program(5,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"radio_big_buttons",{name:"is",hash:{},fn:n.program(7,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"checkbox_big_buttons",{name:"is",hash:{},fn:n.program(7,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"binary_toggle",{name:"is",hash:{},fn:n.program(11,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"dropdown",{name:"is",hash:{},fn:n.program(16,t,0,o,i),inverse:n.noop,data:t}))?r:"")+'\n<div style="clear:both"></div>'},useData:!0,useDepths:!0}),e["place-detail-promotion-bar"]=n({1:function(n,e,a,l,t){return" single-line"},3:function(n,e,a,l,t){return"single-line"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{};return'<section class="promotion clearfix'+(null!=(o=a.if.call(i,null!=e?e.story:e,{name:"if",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t}))?o:"")+'">\n  <div class="support '+(null!=(o=a.if.call(i,null!=e?e.story:e,{name:"if",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t}))?o:"")+'"></div>\n  <div class="social '+(null!=(o=a.if.call(i,null!=e?e.story:e,{name:"if",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t}))?o:"")+'">\n    <ul class="sharing unstyled-list">\n      <li class="share-link share-twitter"><a>Tweet This</a></li>\n      <li class="share-link share-facebook"><a>Recommend on Facebook</a></li>\n    </ul>\n  </div>\n</section>'},useData:!0}),e["place-detail-story-bar-tagline"]=n({1:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{};return'\t<section class="place-story-bar">\n\t\t<span class="btn-story-nav btn-previous-story-nav '+(null!=(o=a.unless.call(i,null!=(o=null!=e?e.story:e)?o.previous:o,{name:"unless",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:"")+'"></span>\n\t\t<span class="story-tagline"><h3>'+n.escapeExpression(n.lambda(null!=(o=null!=e?e.story:e)?o.tagline:o,e))+'</h3></span>\n\t\t<span class="btn-story-nav btn-next-story-nav '+(null!=(o=a.unless.call(i,null!=(o=null!=e?e.story:e)?o.next:o,{name:"unless",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:"")+'"></span>\n\t</section>\n'},2:function(n,e,a,l,t){return"story-nav-disabled"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.story:e,{name:"if",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t}))?o:""},useData:!0}),e["place-detail-story-bar"]=n({1:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{};return'\t<section class="place-story-bar">\n\t\t<span class="btn-story-nav btn-previous-story-nav '+(null!=(o=a.unless.call(i,null!=(o=null!=e?e.story:e)?o.previous:o,{name:"unless",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:"")+'"></span>\n\t\t<span class="story-tagline">\n\t\t\t<h1>\n'+(null!=(o=a.if.call(i,null!=e?e.fullTitle:e,{name:"if",hash:{},fn:n.program(4,t,0),inverse:n.program(6,t,0),data:t}))?o:"")+'\t\t\t</h1>\n\t\t</span>\n\t\t<span class="btn-story-nav btn-next-story-nav '+(null!=(o=a.unless.call(i,null!=(o=null!=e?e.story:e)?o.next:o,{name:"unless",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:"")+'"></span>\n\t</section>\n'},2:function(n,e,a,l,t){return"story-nav-disabled"},4:function(n,e,a,l,t){var o;return"\t\t\t\t\t"+n.escapeExpression((o=null!=(o=a.fullTitle||(null!=e?e.fullTitle:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"fullTitle",hash:{},data:t}):o))+"\n"},6:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.title:e,{name:"if",hash:{},fn:n.program(7,t,0),inverse:n.program(9,t,0),data:t}))?o:""},7:function(n,e,a,l,t){var o;return"\t\t\t\t\t\t"+n.escapeExpression((o=null!=(o=a.title||(null!=e?e.title:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"title",hash:{},data:t}):o))+"\n"},9:function(n,e,a,l,t){var o;return"\t\t\t\t\t\t"+n.escapeExpression((o=null!=(o=a.name||(null!=e?e.name:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"name",hash:{},data:t}):o))+"\n"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.story:e,{name:"if",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t}))?o:""},useData:!0}),e["place-detail-support"]=n({1:function(n,e,a,l,t){return' checked="checked"'},3:function(n,e,a,l,t){var o;return n.escapeExpression((o=null!=(o=a.count||(null!=e?e.count:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"count",hash:{},data:t}):o))},5:function(n,e,a,l,t){return"0"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c='    <form action="#" method="post" class="btn btn-block btn-small user-support">\n      <input type="hidden" name="user_token" value="'+(null!=(i=null!=(i=a.user_token||(null!=e?e.user_token:e))?i:u,o="function"==typeof i?i.call(s,{name:"user_token",hash:{},data:t}):i)?o:"")+'">\n      <input type="hidden" name="visible" value="true">\n      <input type="checkbox" id="support"';return i=null!=(i=a.is_supporting||(null!=e?e.is_supporting:e))?i:u,r={name:"is_supporting",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.is_supporting||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(c+=o),c+'>\n      <label for="support"><span class="support-count">'+(null!=(o=a.if.call(s,null!=e?e.count:e,{name:"if",hash:{},fn:n.program(3,t,0),inverse:n.program(5,t,0),data:t}))?o:"")+"</span> "+n.escapeExpression(n.lambda(null!=(o=null!=e?e.support_config:e)?o.submit_btn_text:o,e))+"</label>\n    </form>\n"},useData:!0}),e["place-detail-survey"]=n({1:function(n,e,a,l,t,o,i){var r,s,u,c=null!=e?e:n.nullContext||{},p=a.helperMissing,m=a.blockHelperMissing,h='          <h4 class="survey-header">\n            '+n.escapeExpression(n.lambda(null!=(r=null!=e?e.responses:e)?r.length:r,e))+"\n            ";return s=null!=(s=a.has_single_response||(null!=e?e.has_single_response:e))?s:p,u={name:"has_single_response",hash:{},fn:n.program(2,t,0,o,i),inverse:n.noop,data:t},r="function"==typeof s?s.call(c,u):s,a.has_single_response||(r=m.call(e,r,u)),null!=r&&(h+=r),h+="\n            ",s=null!=(s=a.has_single_response||(null!=e?e.has_single_response:e))?s:p,u={name:"has_single_response",hash:{},fn:n.noop,inverse:n.program(4,t,0,o,i),data:t},r="function"==typeof s?s.call(c,u):s,a.has_single_response||(r=m.call(e,r,u)),null!=r&&(h+=r),h+="\n          </h4>\n\n"+(null!=(r=a.if.call(c,null!=(r=null!=e?e.survey_config:e)?r.single_submission:r,{name:"if",hash:{},fn:n.program(6,t,0,o,i),inverse:n.program(9,t,0,o,i),data:t}))?r:"")+'\n          <ol class="responses unstyled-list">\n',s=null!=(s=a.responses||(null!=e?e.responses:e))?s:p,u={name:"responses",hash:{},fn:n.program(11,t,0,o,i),inverse:n.noop,data:t},r="function"==typeof s?s.call(c,u):s,a.responses||(r=m.call(e,r,u)),null!=r&&(h+=r),h+"          </ol>\n"},2:function(n,e,a,l,t){var o;return n.escapeExpression(n.lambda(null!=(o=null!=e?e.survey_config:e)?o.response_name:o,e))},4:function(n,e,a,l,t){var o;return n.escapeExpression(n.lambda(null!=(o=null!=e?e.survey_config:e)?o.response_plural_name:o,e))},6:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.user_submitted:e,{name:"if",hash:{},fn:n.noop,inverse:n.program(7,t,0),data:t}))?o:""},7:function(n,e,a,l,t){var o;return'            <a href="#reply" class="reply-link btn btn-small">'+n.escapeExpression(n.lambda(null!=(o=null!=e?e.survey_config:e)?o.form_link_text:o,e))+"</a>\n"},9:function(n,e,a,l,t){var o;return'          <a href="#reply" class="reply-link btn btn-small">'+n.escapeExpression(n.lambda(null!=(o=null!=e?e.survey_config:e)?o.form_link_text:o,e))+"</a>\n"},11:function(n,e,a,l,t,o,i){var r,s,u,c=null!=e?e:n.nullContext||{},p=a.helperMissing,m=n.escapeExpression,h='              <li>\n                <article data-response-id="'+m((s=null!=(s=a.id||(null!=e?e.id:e))?s:p,"function"==typeof s?s.call(c,{name:"id",hash:{},data:t}):s))+'" class="response">\n                  <header class="response-header clearfix">\n                    <strong class="response-submitter">\n'+(null!=(r=a.if.call(c,null!=e?e.submitter:e,{name:"if",hash:{},fn:n.program(12,t,0,o,i),inverse:n.program(14,t,0,o,i),data:t}))?r:"")+'                    </strong>\n                    <span class="response-date">\n                      around '+m((s=null!=(s=a.pretty_created_datetime||(null!=e?e.pretty_created_datetime:e))?s:p,"function"==typeof s?s.call(c,{name:"pretty_created_datetime",hash:{},data:t}):s))+'\n                    </span>\n                    \n                  </header>\n\n                  <div class="response-body">\n';return s=null!=(s=a.items||(null!=e?e.items:e))?s:p,u={name:"items",hash:{},fn:n.program(16,t,0,o,i),inverse:n.noop,data:t},r="function"==typeof s?s.call(c,u):s,a.items||(r=a.blockHelperMissing.call(e,r,u)),null!=r&&(h+=r),h+"                  </div>\n                </article>\n              </li>\n"},12:function(n,e,a,l,t){var o,i=n.lambda,r=n.escapeExpression;return'                        <img src="'+r(i(null!=(o=null!=e?e.submitter:e)?o.avatar_url:o,e))+'" class="avatar" /> '+r(i(null!=(o=null!=e?e.submitter:e)?o.name:o,e))+"\n"},14:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression
+;return"                        \x3c!-- TODO: FIXME: don't hardcode image URL --\x3e\n                        <img src=\""+s((o=null!=(o=a.STATIC_URL||(null!=e?e.STATIC_URL:e))?o:r,"function"==typeof o?o.call(i,{name:"STATIC_URL",hash:{},data:t}):o))+'css/images/user-50.png" class="avatar" /> '+s((o=null!=(o=a.submitter_name||(null!=e?e.submitter_name:e))?o:r,"function"==typeof o?o.call(i,{name:"submitter_name",hash:{},data:t}):o))+"\n"},16:function(n,e,a,l,t,o,i){var r,s,u=null!=e?e:n.nullContext||{},c=a.helperMissing,p=n.escapeExpression;return'                      <div class="response-item response-item-'+p((s=null!=(s=a.name||(null!=e?e.name:e))?s:c,"function"==typeof s?s.call(u,{name:"name",hash:{},data:t}):s))+'" data-cid='+p(n.lambda(null!=i[1]?i[1].cid:i[1],e))+'>\n                        <span class="response-label response-label-'+p((s=null!=(s=a.name||(null!=e?e.name:e))?s:c,"function"==typeof s?s.call(u,{name:"name",hash:{},data:t}):s))+'">'+p((s=null!=(s=a.label||(null!=e?e.label:e))?s:c,"function"==typeof s?s.call(u,{name:"label",hash:{},data:t}):s))+"</span>\n"+(null!=(r=a.if.call(u,null!=i[2]?i[2].isEditingToggled:i[2],{name:"if",hash:{},fn:n.program(17,t,0,o,i),inverse:n.program(19,t,0,o,i),data:t}))?r:"")+"\n"+(null!=(r=a.if.call(u,null!=i[2]?i[2].isEditingToggled:i[2],{name:"if",hash:{},fn:n.program(21,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"                      </div>\n"},17:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'                          <form>\n                            <textarea class="response-value response-value-'+s((o=null!=(o=a.name||(null!=e?e.name:e))?o:r,"function"==typeof o?o.call(i,{name:"name",hash:{},data:t}):o))+'" name='+s((o=null!=(o=a.name||(null!=e?e.name:e))?o:r,"function"==typeof o?o.call(i,{name:"name",hash:{},data:t}):o))+">"+s(a.nlToBr.call(i,null!=e?e.value:e,{name:"nlToBr",hash:{},data:t}))+"</textarea>\n                          </form>\n"},19:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=n.escapeExpression;return'                          <p class="response-value response-value-'+r((o=null!=(o=a.name||(null!=e?e.name:e))?o:a.helperMissing,"function"==typeof o?o.call(i,{name:"name",hash:{},data:t}):o))+'">'+r(a.nlToBr.call(i,null!=e?e.value:e,{name:"nlToBr",hash:{},data:t}))+"</p>\n"},21:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'                          <div id="delete-response-btn-'+s((o=null!=(o=a.name||(null!=e?e.name:e))?o:r,"function"==typeof o?o.call(i,{name:"name",hash:{},data:t}):o))+'" class="btn btn-small editor-btn btn-delete delete-response-btn">\n                            <i class="fa fa-times" aria-hidden="true"></i> Remove comment\n                          </div>\n                          <div id="update-response-btn-'+s((o=null!=(o=a.name||(null!=e?e.name:e))?o:r,"function"==typeof o?o.call(i,{name:"name",hash:{},data:t}):o))+'" class="btn btn-small editor-btn btn-update update-response-btn faded">\n                            <i class="fa fa-floppy-o" aria-hidden="true"></i> Update comment\n                          </div>\n'},23:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.user_submitted:e,{name:"if",hash:{},fn:n.program(24,t,0),inverse:n.program(26,t,0),data:t}))?o:""},24:function(n,e,a,l,t){var o;return null!=(o=n.invokePartial(l["place-survey-response"],e,{name:"place-survey-response",data:t,indent:"              ",helpers:a,partials:l,decorators:n.decorators}))?o:""},26:function(n,e,a,l,t){var o;return null!=(o=a.unless.call(null!=e?e:n.nullContext||{},null!=e?e.isEditingToggled:e,{name:"unless",hash:{},fn:n.program(27,t,0),inverse:n.noop,data:t}))?o:""},27:function(n,e,a,l,t){var o;return null!=(o=n.invokePartial(l["place-survey-form"],e,{name:"place-survey-form",data:t,indent:"                ",helpers:a,partials:l,decorators:n.decorators}))?o:""},29:function(n,e,a,l,t){var o;return null!=(o=a.unless.call(null!=e?e:n.nullContext||{},null!=e?e.isEditingToggled:e,{name:"unless",hash:{},fn:n.program(30,t,0),inverse:n.noop,data:t}))?o:""},30:function(n,e,a,l,t){var o;return null!=(o=n.invokePartial(l["place-survey-form"],e,{name:"place-survey-form",data:t,indent:"              ",helpers:a,partials:l,decorators:n.decorators}))?o:""},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t,o,i){var r;return(null!=(r=a.blockHelperMissing.call(e,n.lambda(null!=(r=null!=e?e.survey_config:e)?r.show_responses:r,e),{name:"survey_config.show_responses",hash:{},fn:n.program(1,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.if.call(null!=e?e:n.nullContext||{},null!=(r=null!=e?e.survey_config:e)?r.single_submission:r,{name:"if",hash:{},fn:n.program(23,t,0,o,i),inverse:n.program(29,t,0,o,i),data:t}))?r:"")},usePartial:!0,useData:!0,useDepths:!0}),e["place-detail"]=n({1:function(n,e,a,l,t){return" editing-toggled"},3:function(n,e,a,l,t){var o;return'      <input type="text" \n             class="place-value place-value-title" \n             name="title" \n             value="'+(null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.fullTitle:e,{name:"if",hash:{},fn:n.program(4,t,0),inverse:n.program(6,t,0),data:t}))?o:"")+'" />\n'},4:function(n,e,a,l,t){var o;return n.escapeExpression((o=null!=(o=a.fullTitle||(null!=e?e.fullTitle:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"fullTitle",hash:{},data:t}):o))},6:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.title:e,{name:"if",hash:{},fn:n.program(7,t,0),inverse:n.program(9,t,0),data:t}))?o:""},7:function(n,e,a,l,t){var o;return n.escapeExpression((o=null!=(o=a.title||(null!=e?e.title:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"title",hash:{},data:t}):o))},9:function(n,e,a,l,t){var o;return n.escapeExpression((o=null!=(o=a.name||(null!=e?e.name:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"name",hash:{},data:t}):o))},11:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{};return'      <h1 class="place-header-title '+(null!=(o=a.if.call(i,null!=e?e.story:e,{name:"if",hash:{},fn:n.program(12,t,0),inverse:n.noop,data:t}))?o:"")+'">\n'+(null!=(o=a.if.call(i,null!=e?e.fullTitle:e,{name:"if",hash:{},fn:n.program(14,t,0),inverse:n.program(16,t,0),data:t}))?o:"")+"      </h1>\n"},12:function(n,e,a,l,t){return"is-visuallyhidden"},14:function(n,e,a,l,t){var o;return"          "+n.escapeExpression((o=null!=(o=a.fullTitle||(null!=e?e.fullTitle:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"fullTitle",hash:{},data:t}):o))+"\n"},16:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.title:e,{name:"if",hash:{},fn:n.program(17,t,0),inverse:n.program(19,t,0),data:t}))?o:""},17:function(n,e,a,l,t){var o;return"            "+n.escapeExpression((o=null!=(o=a.title||(null!=e?e.title:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"title",hash:{},data:t}):o))+"\n"},19:function(n,e,a,l,t){var o;return"            "+n.escapeExpression((o=null!=(o=a.name||(null!=e?e.name:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"name",hash:{},data:t}):o))+"\n"},21:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{},s=n.escapeExpression,u=a.helperMissing;return'      <strong class="point-submitter">\n'+(null!=(o=a.if.call(r,null!=(o=null!=e?e.submitter:e)?o.avatar_url:o,{name:"if",hash:{},fn:n.program(22,t,0),inverse:n.program(24,t,0),data:t}))?o:"")+(null!=(o=a.if.call(r,null!=(o=null!=e?e.submitter:e)?o.name:o,{name:"if",hash:{},fn:n.program(26,t,0),inverse:n.program(28,t,0),data:t}))?o:"")+"      </strong> "+s(a.action_text.call(r,{name:"action_text",hash:{},data:t}))+" this "+s(a.place_type_label.call(r,null!=e?e.location_type:e,{name:"place_type_label",hash:{},data:t}))+"\n\n"+(null!=(o=a.if.call(r,null!=e?e.region:e,{name:"if",hash:{},fn:n.program(33,t,0),inverse:n.noop,data:t}))?o:"")+'\n      <time datetime="'+s((i=null!=(i=a.created_datetime||(null!=e?e.created_datetime:e))?i:u,"function"==typeof i?i.call(r,{name:"created_datetime",hash:{},data:t}):i))+'" \n            class="response-date">\n        <a class="response-date-link" href="/'+s((i=null!=(i=a.datasetSlug||(null!=e?e.datasetSlug:e))?i:u,"function"==typeof i?i.call(r,{name:"datasetSlug",hash:{},data:t}):i))+"/"+s((i=null!=(i=a.id||(null!=e?e.id:e))?i:u,"function"==typeof i?i.call(r,{name:"id",hash:{},data:t}):i))+'">'+s(a.fromnow.call(r,null!=e?e.created_datetime:e,{name:"fromnow",hash:{},data:t}))+'</a>\n      </time>\n\n      <span class="survey-count">'+s(a.survey_count.call(r,{name:"survey_count",hash:{},data:t}))+" "+s(a.survey_label_by_count.call(r,{name:"survey_label_by_count",hash:{},data:t}))+"</span>\n"},22:function(n,e,a,l,t){var o;return'          <img src="'+n.escapeExpression(n.lambda(null!=(o=null!=e?e.submitter:e)?o.avatar_url:o,e))+'" \n               class="avatar" />\n'},24:function(n,e,a,l,t){var o;return'          <img src="'+n.escapeExpression((o=null!=(o=a.STATIC_URL||(null!=e?e.STATIC_URL:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"STATIC_URL",hash:{},data:t}):o))+'css/images/user-50.png" \n               class="avatar" />\n'},26:function(n,e,a,l,t){var o;return"          "+n.escapeExpression(n.lambda(null!=(o=null!=e?e.submitter:e)?o.name:o,e))+"\n"},28:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.submitter_name:e,{name:"if",hash:{},fn:n.program(29,t,0),inverse:n.program(31,t,0),data:t}))?o:""},29:function(n,e,a,l,t){var o;return"            "+n.escapeExpression((o=null!=(o=a.submitter_name||(null!=e?e.submitter_name:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"submitter_name",hash:{},data:t}):o))+"\n"},31:function(n,e,a,l,t){var o;return"            "+n.escapeExpression((o=null!=(o=a.anonymous_name||(null!=e?e.anonymous_name:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"anonymous_name",hash:{},data:t}):o))+"\n"},33:function(n,e,a,l,t){var o;return"        in "+n.escapeExpression((o=null!=(o=a.region||(null!=e?e.region:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"region",hash:{},data:t}):o))+"\n"},35:function(n,e,a,l,t){var o;return'    <a rel="internal" \n       href="/'+(null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e["url-title"]:e,{name:"if",hash:{},fn:n.program(36,t,0),inverse:n.program(38,t,0),data:t}))?o:"")+'" \n       class="view-on-map-btn btn btn-small">在地圖上查看</a>\n'},36:function(n,e,a,l,t){var o;return n.escapeExpression((o=null!=(o=a["url-title"]||(null!=e?e["url-title"]:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"url-title",hash:{},data:t}):o))},38:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return s((o=null!=(o=a.datasetSlug||(null!=e?e.datasetSlug:e))?o:r,"function"==typeof o?o.call(i,{name:"datasetSlug",hash:{},data:t}):o))+"/"+s((o=null!=(o=a.id||(null!=e?e.id:e))?o:r,"function"==typeof o?o.call(i,{name:"id",hash:{},data:t}):o))},40:function(n,e,a,l,t){var o;return null!=(o=a.each.call(null!=e?e:n.nullContext||{},null!=e?e.attachments:e,{name:"each",hash:{},fn:n.program(41,t,0),inverse:n.noop,data:t}))?o:""},41:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'        <div class="place-item place-item-attachment place-attachment-'+s((o=null!=(o=a.name||(null!=e?e.name:e))?o:r,"function"==typeof o?o.call(i,{name:"name",hash:{},data:t}):o))+'">\n          <img src="'+s((o=null!=(o=a.file||(null!=e?e.file:e))?o:r,"function"==typeof o?o.call(i,{name:"file",hash:{},data:t}):o))+'" \n               class="place-value place-value-'+s((o=null!=(o=a.name||(null!=e?e.name:e))?o:r,"function"==typeof o?o.call(i,{name:"name",hash:{},data:t}):o))+'" \n               alt="'+s((o=null!=(o=a.name||(null!=e?e.name:e))?o:r,"function"==typeof o?o.call(i,{name:"name",hash:{},data:t}):o))+'" />\n        </div>           \n'},43:function(n,e,a,l,t,o,i){var r,s,u=null!=e?e:n.nullContext||{};return'      <div class="place-item place-item-'+n.escapeExpression((s=null!=(s=a.name||(null!=e?e.name:e))?s:a.helperMissing,"function"==typeof s?s.call(u,{name:"name",hash:{},data:t}):s))+'">\n'+(null!=(r=a.if.call(u,null!=i[1]?i[1].isEditingToggled:i[1],{name:"if",hash:{},fn:n.program(44,t,0,o,i),inverse:n.program(46,t,0,o,i),data:t}))?r:"")+'        <div style="clear:both"></div>\n      </div>\n'},44:function(n,e,a,l,t){var o;return null!=(o=n.invokePartial(l["place-form-field-types"],e,{name:"place-form-field-types",data:t,indent:"          ",helpers:a,partials:l,decorators:n.decorators}))?o:""},46:function(n,e,a,l,t){var o;return null!=(o=a.unless.call(null!=e?e:n.nullContext||{},null!=e?e.hideFromDetailView:e,{name:"unless",hash:{},fn:n.program(47,t,0),inverse:n.noop,data:t}))?o:""},47:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{},s=a.helperMissing,u=n.escapeExpression;return'            <div class="place-label place-label-'+u((i=null!=(i=a.name||(null!=e?e.name:e))?i:s,"function"==typeof i?i.call(r,{name:"name",hash:{},data:t}):i))+'">'+u((i=null!=(i=a.display_prompt||(null!=e?e.display_prompt:e))?i:s,"function"==typeof i?i.call(r,{name:"display_prompt",hash:{},data:t}):i))+"</div>\n"+(null!=(o=n.invokePartial(l["place-detail-field-types"],e,{name:"place-detail-field-types",data:t,indent:"            ",helpers:a,partials:l,decorators:n.decorators}))?o:"")},49:function(n,e,a,l,t){return'  <section class="survey" \n           id="survey"></section>\n'},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t,o,i){var r,s=null!=e?e:n.nullContext||{};return"\x3c!-- Template for managing the place detail view, which might be rendered in\n  regular mode or in editor mode. --\x3e\n\n"+(null!=(r=n.invokePartial(l["place-detail-edit-bar"],e,{name:"place-detail-edit-bar",data:t,helpers:a,partials:l,decorators:n.decorators}))?r:"")+'\n<div class="place-detail-spacer'+(null!=(r=a.if.call(s,null!=e?e.isEditable:e,{name:"if",hash:{},fn:n.program(1,t,0,o,i),inverse:n.noop,data:t}))?r:"")+'"></div>\n\n'+(null!=(r=n.invokePartial(l["place-detail-story-bar"],e,{name:"place-detail-story-bar",data:t,helpers:a,partials:l,decorators:n.decorators}))?r:"")+"\n"+(null!=(r=n.invokePartial(l["place-detail-promotion-bar"],e,{name:"place-detail-promotion-bar",data:t,helpers:a,partials:l,decorators:n.decorators}))?r:"")+'\n<header class="place-header clearfix">\n'+(null!=(r=n.invokePartial(l["location-string"],null!=e?e.location:e,{name:"location-string",data:t,indent:"  ",helpers:a,partials:l,decorators:n.decorators}))?r:"")+'  <form id="update-place-model-title-form">\n'+(null!=(r=a.if.call(s,null!=e?e.isEditingToggled:e,{name:"if",hash:{},fn:n.program(3,t,0,o,i),inverse:n.program(11,t,0,o,i),data:t}))?r:"")+'  </form>\n  <span class="place-submission-details">\n'+(null!=(r=a.is_not.call(s,null!=e?e.showMetadata:e,!1,{name:"is_not",hash:{},fn:n.program(21,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.if.call(s,null!=e?e.survey_config:e,{name:"if",hash:{},fn:n.noop,inverse:n.program(35,t,0,o,i),data:t}))?r:"")+'\n  </span>\n</header>\n\n<section class="place-items">\n  <form id="update-place-model-form">\n'+(null!=(r=a.unless.call(s,null!=e?e.suppressAttachments:e,{name:"unless",hash:{},fn:n.program(40,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.each.call(s,null!=e?e.fields:e,{name:"each",hash:{},fn:n.program(43,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"  </form>\n</section>\n\n"+(null!=(r=a.if.call(s,null!=e?e.survey_config:e,{name:"if",hash:{},fn:n.program(49,t,0,o,i),inverse:n.noop,data:t}))?r:"")},usePartial:!0,useData:!0,useDepths:!0}),e["place-form-category-buttons"]=n({1:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.isRenderable:e,{name:"if",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:""},2:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{},s=a.helperMissing,u=n.escapeExpression;return'      <div class="category-btn-outer-container">\n\t\t\t\t<input type="radio" \n\t\t\t\t       id='+u((i=null!=(i=a.category||(null!=e?e.category:e))?i:s,"function"==typeof i?i.call(r,{name:"category",hash:{},data:t}):i))+' \n               name="category-radio-btns"\n\t\t\t\t       class="category-btn" />\n\t\t\t\t<label for='+u((i=null!=(i=a.category||(null!=e?e.category:e))?i:s,"function"==typeof i?i.call(r,{name:"category",hash:{},data:t}):i))+'\n\t\t\t\t       class="category-btn-container">\n\t\t\t\t  <span class="category-icon">\n\t\t\t\t    <img src='+u((i=null!=(i=a.icon_url||(null!=e?e.icon_url:e))?i:s,"function"==typeof i?i.call(r,{name:"icon_url",hash:{},data:t}):i))+' />\n\t\t\t\t  </span>\n          <span class="category-label-container">\n\t\t\t\t    <span class="category-label">'+u((i=null!=(i=a.label||(null!=e?e.label:e))?i:s,"function"==typeof i?i.call(r,{name:"label",hash:{},data:t}):i))+'</span>\n            <span class="expansion-icon-container"></span>\n          </span>\n        </label>\n'+(null!=(o=a.if.call(r,null!=e?e.horizontal_rule:e,{name:"if",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t}))?o:"")+"      </div>\n"},3:function(n,e,a,l,t){return"          <hr />\n"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o;return'\x3c!-- Template for dynamic form category selection buttons --\x3e\n\n<div id="category-btns">\n'+(null!=(o=a.each.call(null!=e?e:n.nullContext||{},null!=(o=null!=e?e.placeConfig:e)?o.place_detail:o,{name:"each",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t}))?o:"")+'</div>\n<div style="clear:both"></div>'},useData:!0}),e["place-form-field-types"]=n({1:function(n,e,a,l,t){return"  <hr />\n"},3:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=n.escapeExpression,p='  <label for="place-'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'">\n    '+c((i=null!=(i=a.prompt||(null!=e?e.prompt:e))?i:u,"function"==typeof i?i.call(s,{name:"prompt",hash:{},data:t}):i))+" \n";return i=null!=(i=a.optional||(null!=e?e.optional:e))?i:u,r={name:"optional",hash:{},fn:n.program(4,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.optional||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(p+=o),p+"    "+c((i=null!=(i=a.admin_msg||(null!=e?e.admin_msg:e))?i:u,"function"==typeof i?i.call(s,{name:"admin_msg",hash:{},data:t}):i))+"\n  </label>\n"},4:function(n,e,a,l,t){return"      <small>(optional)</small>\n"},6:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=n.escapeExpression,p='  <div class="geocoding-enabled">\n    <div id="geocode-address-place-bar" class="clearfix">\n      <p class="error is-hidden"></p>\n      <label class="geocode-address-label" for="geocode-address-field"></label>\n      <div class="geocode-spinner is-hidden"></div>\n      <input name="'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'" \n             ';return i=null!=(i=a.attrs||(null!=e?e.attrs:e))?i:u,r={name:"attrs",hash:{},fn:n.program(7,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.attrs||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(p+=o),p+' \n             type="text" id="geocode-address-field" \n             class="geocode-address-field"\n             value="'+c((i=null!=(i=a.content||(null!=e?e.content:e))?i:u,"function"==typeof i?i.call(s,{name:"content",hash:{},data:t}):i))+'">\n    </div>\n  </div>\n'},7:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return" "+s((o=null!=(o=a.key||(null!=e?e.key:e))?o:r,"function"==typeof o?o.call(i,{name:"key",hash:{},data:t}):o))+'="'+s((o=null!=(o=a.value||(null!=e?e.value:e))?o:r,"function"==typeof o?o.call(i,{name:"value",hash:{},data:t}):o))+'"'},9:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=n.escapeExpression,p='  <input id="datetimepicker" \n         name="'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'" \n         type="text"\n         '+(null!=(o=a.if.call(s,null!=e?e.isAutocomplete:e,{name:"if",hash:{},fn:n.program(10,t,0),inverse:n.noop,data:t}))?o:"")+"\n         ";return i=null!=(i=a.optional||(null!=e?e.optional:e))?i:u,r={name:"optional",hash:{},fn:n.noop,inverse:n.program(12,t,0),data:t},o="function"==typeof i?i.call(s,r):i,a.optional||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(p+=o),p+"\n         value="+c(a.nlToBr.call(s,null!=e?e.content:e,{name:"nlToBr",hash:{},data:t}))+">\n"},10:function(n,e,a,l,t){return'class="is-autocomplete"'},12:function(n,e,a,l,t){return"required"},14:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=n.escapeExpression,m=a.blockHelperMissing,h='  <input id="place-'+p((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,typeof i===c?i.call(s,{name:"name",hash:{},data:t}):i))+'" \n         '+(null!=(o=a.if.call(s,null!=e?e.isAutocomplete:e,{name:"if",hash:{},fn:n.program(10,t,0),inverse:n.noop,data:t}))?o:"")+'\n         name="'+p((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,typeof i===c?i.call(s,{name:"name",hash:{},data:t}):i))+'" \n         type="'+p((i=null!=(i=a.type||(null!=e?e.type:e))?i:u,typeof i===c?i.call(s,{name:"type",hash:{},data:t}):i))+'" \n         ';return i=null!=(i=a.attrs||(null!=e?e.attrs:e))?i:u,r={name:"attrs",hash:{},fn:n.program(7,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a.attrs||(o=m.call(e,o,r)),null!=o&&(h+=o),h+=" \n         "+(null!=(o=a.unless.call(s,null!=e?e.autocomplete:e,{name:"unless",hash:{},fn:n.program(15,t,0),inverse:n.noop,data:t}))?o:"")+"\n         ",i=null!=(i=a.optional||(null!=e?e.optional:e))?i:u,r={name:"optional",hash:{},fn:n.noop,inverse:n.program(12,t,0),data:t},o=typeof i===c?i.call(s,r):i,a.optional||(o=m.call(e,o,r)),null!=o&&(h+=o),h+'\n         value="'+p(a.nlToBr.call(s,null!=e?e.content:e,{name:"nlToBr",hash:{},data:t}))+'">\n'},15:function(n,e,a,l,t){var o;return' placeholder="'+n.escapeExpression((o=null!=(o=a.placeholder||(null!=e?e.placeholder:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"placeholder",hash:{},data:t}):o))+'"'},17:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=n.escapeExpression,p=a.blockHelperMissing,m='  <textarea id="place-'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'" \n            '+(null!=(o=a.if.call(s,null!=e?e.isAutocomplete:e,{name:"if",hash:{},fn:n.program(10,t,0),inverse:n.noop,data:t}))?o:"")+'\n            name="'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'" \n            ';return i=null!=(i=a.attrs||(null!=e?e.attrs:e))?i:u,r={name:"attrs",hash:{},fn:n.program(7,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.attrs||(o=p.call(e,o,r)),null!=o&&(m+=o),m+=" \n            "+(null!=(o=a.unless.call(s,null!=e?e.autocomplete:e,{name:"unless",hash:{},fn:n.program(15,t,0),inverse:n.noop,data:t}))?o:"")+"\n            ",i=null!=(i=a.optional||(null!=e?e.optional:e))?i:u,r={name:"optional",hash:{},fn:n.noop,inverse:n.program(12,t,0),data:t},o="function"==typeof i?i.call(s,r):i,a.optional||(o=p.call(e,o,r)),null!=o&&(m+=o),m+">"+c(a.nlToBr.call(s,null!=e?e.content:e,{name:"nlToBr",hash:{},data:t}))+"</textarea>\n"},19:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=n.escapeExpression,m=a.blockHelperMissing,h='  <div id="place-'+p((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,typeof i===c?i.call(s,{name:"name",hash:{},data:t}):i))+'" \n       class="rich-text-field'+(null!=(o=a.if.call(s,null!=e?e.isAutocomplete:e,{name:"if",hash:{},fn:n.program(20,t,0),inverse:n.noop,data:t}))?o:"")+'"\n       name="'+p((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,typeof i===c?i.call(s,{name:"name",hash:{},data:t}):i))+'" \n       ';return i=null!=(i=a.attrs||(null!=e?e.attrs:e))?i:u,r={name:"attrs",hash:{},fn:n.program(7,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a.attrs||(o=m.call(e,o,r)),null!=o&&(h+=o),h+=" \n       "+(null!=(o=a.unless.call(s,null!=e?e.autcompleteValue:e,{name:"unless",hash:{},fn:n.program(15,t,0),inverse:n.noop,data:t}))?o:"")+"\n       ",i=null!=(i=a.optional||(null!=e?e.optional:e))?i:u,r={name:"optional",hash:{},fn:n.noop,inverse:n.program(12,t,0),data:t},o=typeof i===c?i.call(s,r):i,a.optional||(o=m.call(e,o,r)),null!=o&&(h+=o),h+">"+(null!=(i=null!=(i=a.content||(null!=e?e.content:e))?i:u,o=typeof i===c?i.call(s,{name:"content",hash:{},data:t}):i)?o:"")+"</div>\n"},20:function(n,e,a,l,t){return" is-autocomplete"},22:function(n,e,a,l,t,o,i){var r;return"  <div>\n"+(null!=(r=a.each.call(null!=e?e:n.nullContext||{},null!=e?e.content:e,{name:"each",hash:{},fn:n.program(23,t,0,o,i),inverse:n.noop,data:t}))?r:"")+'  </div>\n  <div style="clear:both"></div>\n'},23:function(n,e,a,l,t,o,i){var r,s,u=n.lambda,c=n.escapeExpression,p=null!=e?e:n.nullContext||{},m=a.helperMissing;return'        <input type="radio" \n               name='+c(u(null!=i[1]?i[1].name:i[1],e))+" \n               id="+c((s=null!=(s=a.value||(null!=e?e.value:e))?s:m,"function"==typeof s?s.call(p,{name:"value",hash:{},data:t}):s))+" value="+c((s=null!=(s=a.value||(null!=e?e.value:e))?s:m,"function"==typeof s?s.call(p,{name:"value",hash:{},data:t}):s))+' \n               class="big-btn" \n               '+(null!=(r=a.blockHelperMissing.call(e,u(null!=i[1]?i[1].optional:i[1],e),{name:"../optional",hash:{},fn:n.noop,inverse:n.program(12,t,0,o,i),data:t}))?r:"")+"\n               "+(null!=(r=a.if.call(p,null!=e?e.selected:e,{name:"if",hash:{},fn:n.program(24,t,0,o,i),inverse:n.noop,data:t}))?r:"")+">\n        <label for="+c((s=null!=(s=a.value||(null!=e?e.value:e))?s:m,"function"==typeof s?s.call(p,{name:"value",hash:{},data:t}):s))+'\n               class="'+(null!=(r=a.ifAnd.call(p,null!=i[1]?i[1].isAutocomplete:i[1],null!=e?e.selected:e,{name:"ifAnd",hash:{},fn:n.program(26,t,0,o,i),inverse:n.noop,data:t}))?r:"")+'">'+c((s=null!=(s=a.label||(null!=e?e.label:e))?s:m,"function"==typeof s?s.call(p,{name:"label",hash:{},data:t}):s))+"</label>\n"},24:function(n,e,a,l,t){return"checked"},26:function(n,e,a,l,t){return"is-autocomplete"},28:function(n,e,a,l,t,o,i){var r;return'  <div class="checkbox-group">\n'+(null!=(r=a.each.call(null!=e?e:n.nullContext||{},null!=e?e.content:e,{name:"each",hash:{},fn:n.program(29,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"  </div>\n"},29:function(n,e,a,l,t,o,i){var r,s,u=n.escapeExpression,c=null!=e?e:n.nullContext||{},p=a.helperMissing;return'      <input type="checkbox" \n             name='+u(n.lambda(null!=i[1]?i[1].name:i[1],e))+" \n             id="+u((s=null!=(s=a.value||(null!=e?e.value:e))?s:p,"function"==typeof s?s.call(c,{name:"value",hash:{},data:t}):s))+" \n             value="+u((s=null!=(s=a.value||(null!=e?e.value:e))?s:p,"function"==typeof s?s.call(c,{name:"value",hash:{},data:t}):s))+' \n             class="big-btn"\n             '+(null!=(r=a.if.call(c,null!=e?e.selected:e,{name:"if",hash:{},fn:n.program(24,t,0,o,i),inverse:n.noop,data:t}))?r:"")+">\n      <label for="+u((s=null!=(s=a.value||(null!=e?e.value:e))?s:p,"function"==typeof s?s.call(c,{name:"value",hash:{},data:t}):s))+'\n             class="'+(null!=(r=a.ifAnd.call(c,null!=i[1]?i[1].isAutocomplete:i[1],null!=e?e.selected:e,{name:"ifAnd",hash:{},fn:n.program(26,t,0,o,i),inverse:n.noop,data:t}))?r:"")+'">'+u((s=null!=(s=a.label||(null!=e?e.label:e))?s:p,"function"==typeof s?s.call(c,{name:"label",hash:{},data:t}):s))+"</label>\n"},31:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{},s=a.helperMissing,u=n.escapeExpression;return'  <div>\n    <input type="checkbox" \n           name='+u((i=null!=(i=a.name||(null!=e?e.name:e))?i:s,"function"==typeof i?i.call(r,{name:"name",hash:{},data:t}):i))+" \n           id="+u((i=null!=(i=a.name||(null!=e?e.name:e))?i:s,"function"==typeof i?i.call(r,{name:"name",hash:{},data:t}):i))+" \n           value="+(null!=(o=a.if.call(r,null!=(o=null!=e?e.content:e)?o.selected:o,{name:"if",hash:{},fn:n.program(32,t,0),inverse:n.program(34,t,0),data:t}))?o:"")+"           data-alt-value="+(null!=(o=a.if.call(r,null!=(o=null!=e?e.content:e)?o.selected:o,{name:"if",hash:{},fn:n.program(36,t,0),inverse:n.program(38,t,0),data:t}))?o:"")+'           data-input-type="binary_toggle" \n           class="big-btn"\n           '+(null!=(o=a.if.call(r,null!=(o=null!=e?e.content:e)?o.selected:o,{name:"if",hash:{},fn:n.program(24,t,0),inverse:n.noop,data:t}))?o:"")+">\n    <label for="+u((i=null!=(i=a.name||(null!=e?e.name:e))?i:s,"function"==typeof i?i.call(r,{name:"name",hash:{},data:t}):i))+"\n           "+(null!=(o=a.ifAnd.call(r,null!=e?e.isAutocomplete:e,null!=(o=null!=e?e.content:e)?o.selected:o,{name:"ifAnd",hash:{},fn:n.program(10,t,0),inverse:n.noop,data:t}))?o:"")+"\n           data-alt-label="+(null!=(o=a.if.call(r,null!=(o=null!=e?e.content:e)?o.selected:o,{name:"if",hash:{},fn:n.program(40,t,0),inverse:n.program(42,t,0),data:t}))?o:"")+">"+(null!=(o=a.if.call(r,null!=(o=null!=e?e.content:e)?o.selected:o,{name:"if",hash:{},fn:n.program(44,t,0),inverse:n.program(46,t,0),data:t}))?o:"")+"</label>\n  </div>\n"},32:function(n,e,a,l,t){var o;return"\n                   "+n.escapeExpression(n.lambda(null!=(o=null!=e?e.content:e)?o.selectedValue:o,e))+"\n"},34:function(n,e,a,l,t){var o;return"                   "+n.escapeExpression(n.lambda(null!=(o=null!=e?e.content:e)?o.unselectedValue:o,e))+"\n"},36:function(n,e,a,l,t){var o;return"\n                            "+n.escapeExpression(n.lambda(null!=(o=null!=e?e.content:e)?o.unselectedValue:o,e))+"\n"},38:function(n,e,a,l,t){var o;return"                            "+n.escapeExpression(n.lambda(null!=(o=null!=e?e.content:e)?o.selectedValue:o,e))+"\n"},40:function(n,e,a,l,t){var o;return"\n                            "+n.escapeExpression(n.lambda(null!=(o=null!=e?e.content:e)?o.unselectedLabel:o,e))+"\n"},42:function(n,e,a,l,t){var o;return"                            "+n.escapeExpression(n.lambda(null!=(o=null!=e?e.content:e)?o.selectedLabel:o,e))+"\n                          "},44:function(n,e,a,l,t){var o;return n.escapeExpression(n.lambda(null!=(o=null!=e?e.content:e)?o.selectedLabel:o,e))},46:function(n,e,a,l,t){var o;return n.escapeExpression(n.lambda(null!=(o=null!=e?e.content:e)?o.unselectedLabel:o,e))},48:function(n,e,a,l,t,o,i){var r,s,u,c=null!=e?e:n.nullContext||{},p="  <div>\n    <select name="+n.escapeExpression(n.lambda(null!=i[1]?i[1].name:i[1],e))+" \n            ";return s=null!=(s=a.optional||(null!=e?e.optional:e))?s:a.helperMissing,u={name:"optional",hash:{},fn:n.noop,inverse:n.program(12,t,0,o,i),data:t},r="function"==typeof s?s.call(c,u):s,a.optional||(r=a.blockHelperMissing.call(e,r,u)),null!=r&&(p+=r),p+"\n            "+(null!=(r=a.if.call(c,null!=e?e.isAutocomplete:e,{name:"if",hash:{},fn:n.program(10,t,0,o,i),inverse:n.noop,data:t}))?r:"")+'>\n      <option value="" name="no_response">Select...</option>\n'+(null!=(r=a.each.call(c,null!=e?e.content:e,{name:"each",hash:{},fn:n.program(49,t,0,o,i),inverse:n.noop,data:t}))?r:"")+'    </select>\n  </div>\n  <div style="clear:both"></div>\n'},49:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{},s=a.helperMissing,u=n.escapeExpression;return"          <option value="+u((i=null!=(i=a.value||(null!=e?e.value:e))?i:s,"function"==typeof i?i.call(r,{name:"value",hash:{},data:t}):i))+" "+(null!=(o=a.if.call(r,null!=e?e.selected:e,{name:"if",hash:{},fn:n.program(50,t,0),inverse:n.noop,data:t
+}))?o:"")+">"+u((i=null!=(i=a.label||(null!=e?e.label:e))?i:s,"function"==typeof i?i.call(r,{name:"label",hash:{},data:t}):i))+"</option>\n"},50:function(n,e,a,l,t){return"selected"},52:function(n,e,a,l,t,o,i){var r,s,u,c=n.lambda,p=n.escapeExpression,m=null!=e?e:n.nullContext||{},h='  <div>\n    <select id="dropdown-autocomplete-container-'+p(c(null!=i[1]?i[1].name:i[1],e))+'"\n            class="dropdown-autocomplete-container"\n            name='+p(c(null!=i[1]?i[1].name:i[1],e))+' \n            data-placeholder="'+p(c(null!=i[1]?i[1].placeholder:i[1],e))+'"\n            ';return s=null!=(s=a.optional||(null!=e?e.optional:e))?s:a.helperMissing,u={name:"optional",hash:{},fn:n.noop,inverse:n.program(12,t,0,o,i),data:t},r="function"==typeof s?s.call(m,u):s,a.optional||(r=a.blockHelperMissing.call(e,r,u)),null!=r&&(h+=r),h+"\n            "+(null!=(r=a.if.call(m,null!=e?e.isAutocomplete:e,{name:"if",hash:{},fn:n.program(10,t,0,o,i),inverse:n.noop,data:t}))?r:"")+'>\n      <option value="" name="no_response"></option>\n'+(null!=(r=a.each.call(m,null!=e?e.content:e,{name:"each",hash:{},fn:n.program(49,t,0,o,i),inverse:n.noop,data:t}))?r:"")+'    </select>\n  </div>\n  <div style="clear:both"></div>\n'},54:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=n.escapeExpression,m=a.blockHelperMissing,h='  <div class="url-readout-container">\n    <p class="url-readout-msg">Your post will be accessible at:</p>\n    <p>\n      <span class="url-prefix"></span>\n      <span class="url-readout"></span>\n    </p>\n  </div>\n  <input id="place-'+p((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,typeof i===c?i.call(s,{name:"name",hash:{},data:t}):i))+'" \n         name="'+p((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,typeof i===c?i.call(s,{name:"name",hash:{},data:t}):i))+'" \n         type="text" \n         ';return i=null!=(i=a.attrs||(null!=e?e.attrs:e))?i:u,r={name:"attrs",hash:{},fn:n.program(7,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a.attrs||(o=m.call(e,o,r)),null!=o&&(h+=o),h+=' \n         placeholder="'+p((i=null!=(i=a.placeholder||(null!=e?e.placeholder:e))?i:u,typeof i===c?i.call(s,{name:"placeholder",hash:{},data:t}):i))+'"\n         ',i=null!=(i=a.optional||(null!=e?e.optional:e))?i:u,r={name:"optional",hash:{},fn:n.noop,inverse:n.program(12,t,0),data:t},o=typeof i===c?i.call(s,r):i,a.optional||(o=m.call(e,o,r)),null!=o&&(h+=o),h+"\n         value="+p(a.nlToBr.call(s,null!=e?e.content:e,{name:"nlToBr",hash:{},data:t}))+'>\n  <div class="invalid-msg-container">\n    <div class="invalid-msg duplicate-url hidden"><em>This url is already in use. Please try another.</em></div>\n    <div class="invalid-msg forward-slash hidden"><em>Please choose a url without a forward slash character (/) in it.</em></div>\n  </div>\n'},56:function(n,e,a,l,t,o,i){var r;return'  <div class="geometry-toolbar">\n    <div class="helper-messages">\n      <p class="create-new-geometry-msg"><em>Click on a tool below to draw on the map</em></p>\n      <p class="modify-geometry-msg hidden"><em>Click on a tool below to modify your shape</em></p>\n      <p class="edit-poly-geometry-msg hidden"><em>Drag shape handles to edit your shape, or click on shape handles to delete vertices</em></p>\n      <p class="edit-marker-geometry-msg hidden"><em>Drag your marker to reposition it</em></p>\n      <p class="draw-poly-geometry-start-msg hidden"><em>Click anywhere on the map to start drawing</em></p>\n      <p class="draw-polygon-continue-msg hidden"><em>Click to continue drawing your shape</em></p>\n      <p class="draw-polygon-finish-msg hidden"><em>Click to continue drawing, or click on your first point to complete your shape</em></p>\n      <p class="draw-polyline-continue-msg hidden"><em>Click to continue drawing</em></p>\n      <p class="draw-polyline-finish-msg hidden"><em>Click to continue drawing, or double click to complete your shape</em></p>\n      <p class="draw-marker-geometry-msg hidden"><em>Click anywhere on the map to place your marker</em></p>\n    </div>\n    <div class="geometry-toolbar-create">\n      <button class="create-marker"><img src="/static/css/images/create-marker-icon.svg"><span>Create marker</span></button>\n      <button class="create-polyline"><img src="/static/css/images/create-polyline-icon.svg"><span>Create polyline</span></button>\n      <button class="create-polygon"><img src="/static/css/images/create-polygon-icon.svg"><span>Create polygon</span></button>\n    </div>\n    <div class="geometry-toolbar-edit hidden">\n      <button class="edit-geometry"><img src="/static/css/images/edit-geometry-icon.svg">\n        <span class="edit-geometry-Point hidden">Edit marker</span>\n        <span class="edit-geometry-LineString hidden">Edit polyline</span>\n        <span class="edit-geometry-Polygon hidden">Edit polygon</span>\n      </button>\n      <button class="delete-geometry"><img src="/static/css/images/delete-geometry-icon.svg">\n        <span class="delete-geometry-Point hidden">Delete marker</span>\n        <span class="delete-geometry-LineString hidden">Delete polyline</span>\n        <span class="delete-geometry-Polygon hidden">Delete polygon</span>\n      </button>\n      <button class="colorpicker"><img src="/static/css/images/colorpicker-icon.svg"><span>Change colors</span></button>\n    </div>\n    <div class="clearfix"></div>\n    <div class="geometry-toolbar-icon-field hidden">\n      <p><em>Choose an icon style below</em></p>\n'+(null!=(r=a.each.call(null!=e?e:n.nullContext||{},null!=e?e.content:e,{name:"each",hash:{},fn:n.program(57,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"    </div>\n  </div>\n"},57:function(n,e,a,l,t,o,i){var r,s,u=n.lambda,c=n.escapeExpression,p=null!=e?e:n.nullContext||{},m=a.helperMissing,h="function";return'        <input type="radio" \n               name='+c(u(null!=i[1]?i[1].name:i[1],e))+' \n               id="'+c(u(null!=i[1]?i[1].name:i[1],e))+"-"+c((s=null!=(s=a.index||t&&t.index)?s:m,typeof s===h?s.call(p,{name:"index",hash:{},data:t}):s))+'" \n               value='+c((s=null!=(s=a.url||(null!=e?e.url:e))?s:m,typeof s===h?s.call(p,{name:"url",hash:{},data:t}):s))+" \n               "+(null!=(r=a.if.call(p,null!=e?e.selected:e,{name:"if",hash:{},fn:n.program(24,t,0,o,i),inverse:n.noop,data:t}))?r:"")+'>\n        <label for="'+c(u(null!=i[1]?i[1].name:i[1],e))+"-"+c((s=null!=(s=a.index||t&&t.index)?s:m,typeof s===h?s.call(p,{name:"index",hash:{},data:t}):s))+'"\n               class="'+(null!=(r=a.ifAnd.call(p,null!=i[1]?i[1].isAutocomplete:i[1],null!=e?e.selected:e,{name:"ifAnd",hash:{},fn:n.program(26,t,0,o,i),inverse:n.noop,data:t}))?r:"")+'">'+c((s=null!=(s=a.label||(null!=e?e.label:e))?s:m,typeof s===h?s.call(p,{name:"label",hash:{},data:t}):s))+"\n          <img src="+c((s=null!=(s=a.url||(null!=e?e.url:e))?s:m,typeof s===h?s.call(p,{name:"url",hash:{},data:t}):s))+" />\n        </label>\n"},59:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{};return'  <div class="publish-control-container">\n    <input type="radio" \n           name="published" \n           id="is-published" \n           value="isPublished"\n           class="big-btn" \n           data-helper-msg="is-published-msg"\n           '+(null!=(o=a.is.call(i,null!=(o=null!=e?e.content:e)?o.selected:o,"isPublished",{name:"is",hash:{},fn:n.program(24,t,0),inverse:n.noop,data:t}))?o:"")+'>\n    <label for="is-published">Published</label>\n    <input type="radio" \n           name="published"\n           id="is-not-published" \n           value="isNotPublished"\n           class="big-btn"\n           data-helper-msg="is-not-published-msg"\n           '+(null!=(o=a.is.call(i,null!=(o=null!=e?e.content:e)?o.selected:o,"isNotPublished",{name:"is",hash:{},fn:n.program(24,t,0),inverse:n.noop,data:t}))?o:"")+'>\n    <label for="is-not-published">Not published</label>\n    <div class="clearfix"></div>\n    <div class="helper-messages">\n      <p class="is-published-msg-form"><em>When created, this post will be visible to <strong>all map users</strong></em></p>\n      <p class="is-not-published-msg-form hidden"><em>When created, this post will be visible <strong>only to logged-in editors</strong></em></p>\n      <p class="is-published-msg-editor'+(null!=(o=a.is_not.call(i,null!=(o=null!=e?e.content:e)?o.selected:o,"isPublished",{name:"is_not",hash:{},fn:n.program(60,t,0),inverse:n.noop,data:t}))?o:"")+'"><em>When saved, this post will be visible to <strong>all map users</strong></em></p>\n      <p class="is-not-published-msg-editor'+(null!=(o=a.is_not.call(i,null!=(o=null!=e?e.content:e)?o.selected:o,"isNotPublished",{name:"is_not",hash:{},fn:n.program(60,t,0),inverse:n.noop,data:t}))?o:"")+'"><em>When saved, this post will be visible <strong>only to logged-in editors</strong></em></p>\n    </div>\n  </div>\n'},60:function(n,e,a,l,t){return" hidden"},62:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=n.escapeExpression,m='  <span class="fileinput-container btn btn-small'+(null!=(o=a.if_fileinput_not_supported.call(s,{name:"if_fileinput_not_supported",hash:{},fn:n.program(63,t,0),inverse:n.noop,data:t}))?o:"")+'">\n    <span>'+p((i=null!=(i=a.inputfile_label||(null!=e?e.inputfile_label:e))?i:u,typeof i===c?i.call(s,{name:"inputfile_label",hash:{},data:t}):i))+'</span>\n    <input id="place-'+p((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,typeof i===c?i.call(s,{name:"name",hash:{},data:t}):i))+'" name="'+p((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,typeof i===c?i.call(s,{name:"name",hash:{},data:t}):i))+'" type="'+p((i=null!=(i=a.type||(null!=e?e.type:e))?i:u,typeof i===c?i.call(s,{name:"type",hash:{},data:t}):i))+'" class="shareabouts-file-input"\n           ';return i=null!=(i=a.attrs||(null!=e?e.attrs:e))?i:u,r={name:"attrs",hash:{},fn:n.program(7,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a.attrs||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(m+=o),m+"\n           "+(null!=(o=a.if_fileinput_not_supported.call(s,{name:"if_fileinput_not_supported",hash:{},fn:n.program(65,t,0),inverse:n.noop,data:t}))?o:"")+'>\n  </span>\n  <span class="fileinput-name">\n    '+(null!=(o=a.if_fileinput_not_supported.call(s,{name:"if_fileinput_not_supported",hash:{},fn:n.program(67,t,0),inverse:n.noop,data:t}))?o:"")+"\n  </span>\n"},63:function(n,e,a,l,t){return" btn-disabled"},65:function(n,e,a,l,t){return" disabled"},67:function(n,e,a,l,t){return"Sorry, your browser doesn't support file uploads."},69:function(n,e,a,l,t){var o;return'  <div>\n    <input name="save-place-btn" id="save-place-btn" type="submit" value="'+n.escapeExpression((o=null!=(o=a.label||(null!=e?e.label:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"label",hash:{},data:t}):o))+'" class="btn btn-primary submit-btn">\n  </div>\n'},71:function(n,e,a,l,t){var o;return'  <div class="form-annotation">'+n.escapeExpression((o=null!=(o=a.annotation||(null!=e?e.annotation:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"annotation",hash:{},data:t}):o))+"</div>\n"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t,o,i){var r,s=null!=e?e:n.nullContext||{};return"\x3c!-- Template for all input types, used by both the place form view and the \n  place detail view in editor mode. --\x3e\n\n"+(null!=(r=a.if.call(s,null!=e?e.horizontal_rule:e,{name:"if",hash:{},fn:n.program(1,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.if.call(s,null!=e?e.prompt:e,{name:"if",hash:{},fn:n.program(3,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"geocoding",{name:"is",hash:{},fn:n.program(6,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"datetime",{name:"is",hash:{},fn:n.program(9,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"text",{name:"is",hash:{},fn:n.program(14,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"textarea",{name:"is",hash:{},fn:n.program(17,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"richTextarea",{name:"is",hash:{},fn:n.program(19,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"radio_big_buttons",{name:"is",hash:{},fn:n.program(22,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"checkbox_big_buttons",{name:"is",hash:{},fn:n.program(28,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"binary_toggle",{name:"is",hash:{},fn:n.program(31,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"dropdown",{name:"is",hash:{},fn:n.program(48,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"dropdown-autocomplete",{name:"is",hash:{},fn:n.program(52,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"url-title",{name:"is",hash:{},fn:n.program(54,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"geometryToolbar",{name:"is",hash:{},fn:n.program(56,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"publishControl",{name:"is",hash:{},fn:n.program(59,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"file",{name:"is",hash:{},fn:n.program(62,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"+(null!=(r=a.is.call(s,null!=e?e.type:e,"submit",{name:"is",hash:{},fn:n.program(69,t,0,o,i),inverse:n.noop,data:t}))?r:"")+'<div style="clear:both"></div>\n\n\x3c!-- optional message appearing after the input type --\x3e\n'+(null!=(r=a.if.call(s,null!=e?e.annotation:e,{name:"if",hash:{},fn:n.program(71,t,0,o,i),inverse:n.noop,data:t}))?r:"")+"\n"},useData:!0,useDepths:!0}),e["place-form-fields"]=n({1:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.isRenderable:e,{name:"if",hash:{},fn:n.program(2,t,0),inverse:n.noop,data:t}))?o:""},2:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{};return'      <div class="'+n.escapeExpression((i=null!=(i=a.type||(null!=e?e.type:e))?i:a.helperMissing,"function"==typeof i?i.call(r,{name:"type",hash:{},data:t}):i))+" form-field"+(null!=(o=a.if.call(r,null!=e?e.is_admin:e,{name:"if",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t}))?o:"")+'">\n'+(null!=(o=n.invokePartial(l["place-form-field-types"],e,{name:"place-form-field-types",data:t,indent:"        ",helpers:a,partials:l,decorators:n.decorators}))?o:"")+"      </div>\n"},3:function(n,e,a,l,t){return" is-admin-field"},5:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.isRenderable:e,{name:"if",hash:{},fn:n.program(6,t,0),inverse:n.noop,data:t}))?o:""},6:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{};return'        <div class="'+n.escapeExpression((i=null!=(i=a.type||(null!=e?e.type:e))?i:a.helperMissing,"function"==typeof i?i.call(r,{name:"type",hash:{},data:t}):i))+" form-field"+(null!=(o=a.if.call(r,null!=e?e.is_admin:e,{name:"if",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t}))?o:"")+'">\n'+(null!=(o=n.invokePartial(l["place-form-field-types"],e,{name:"place-form-field-types",data:t,indent:"          ",helpers:a,partials:l,decorators:n.decorators}))?o:"")+"        </div>\n"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{};return'\x3c!-- Template for dynamic form fields --\x3e\n\n<fieldset>\n  <input type="hidden" name="visible" value="on">\n  <input type="hidden" name="user_token" value="'+(null!=(i=null!=(i=a.user_token||(null!=e?e.user_token:e))?i:a.helperMissing,o="function"==typeof i?i.call(r,{name:"user_token",hash:{},data:t}):i)?o:"")+'">\n\n  \x3c!-- selected category fields --\x3e\n'+(null!=(o=a.each.call(r,null!=(o=null!=e?e.selectedCategoryConfig:e)?o.fields:o,{name:"each",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t}))?o:"")+'\n  \x3c!-- common form elements, included at the bottom of all forms --\x3e\n  <div id="common-form-elements">\n'+(null!=(o=a.each.call(r,null!=e?e.commonFormElements:e,{name:"each",hash:{},fn:n.program(5,t,0),inverse:n.noop,data:t}))?o:"")+'  </div>\n\n  <div class="form-spinner"></div>\n</fieldset>'},usePartial:!0,useData:!0}),e["place-form-messages"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o;return'\x3c!-- Template for dynamic form alert/warning messages --\x3e\n\n<p class="drag-marker-msg alert">首先，拖動地圖以設置您的位置</p>\n<p class="drag-marker-warning-msg warning hidden">您似乎尚未設置自己的位置。請拖動地圖以設置您的位置</p>\n<p class="geolocating-msg alert hidden">Setting your location...</p>\n<p class="create-geometry-msg alert hidden">Use the drawing tools to create a shape on the map.</p>\n<p class="no-geometry-warning-msg warning hidden">It looks like you didn\'t create any geometry yet. Use the drawing tools to create a shape on the map.</p>\n\n<p class="form-field">'+n.escapeExpression(n.lambda(null!=(o=null!=e?e.placeConfig:e)?o.help_text:o,e))+"</p>"},useData:!0}),e["place-form"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'\x3c!-- Template for the dynamic form container, with sections for alert/warning\n\tmessages, category selection buttons, and form fields. --\x3e\n\n\x3c!-- alert/warning messages --\x3e\n<div class="place-form-geometry-messages"></div>\n\n\x3c!-- category selection buttons --\x3e\n<div class="place-form-category-buttons"></div>\n\n\x3c!-- form fields --\x3e\n<form id="place-form" class="place-form clearfix"></form>\n'},useData:!0}),e["place-list"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o=null!=e?e:n.nullContext||{},i=n.escapeExpression;return'<div class="place-list-header">\n  <form class="list-search-form">\n    <input id="list-search" name="list_search" type="text" placeholder="Search...">\n    <input name="search" type="submit" value="搜索" class="btn btn-primary submit-btn">\n  </form>\n  <nav class="list-sort-menu"><span class="list-sort-prompt"></span> <a href="#" class="date-sort is-selected">最近的</a> | <a href="#" class="support-sort">最 '+i(a.support_label_plural.call(o,{name:"support_label_plural",hash:{},data:t}))+'</a> | <a href="#" class="survey-sort">最 '+i(a.survey_label_plural.call(o,{name:"survey_label_plural",hash:{},data:t}))+'</a></nav>\n</div>\n\n<ul class="place-list unstyled-list"></ul>'},useData:!0}),e["place-survey-form"]=n({1:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=n.escapeExpression,m=a.blockHelperMissing,h='\n      <div class="survey-field survey-field-'+p((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,typeof i===c?i.call(s,{name:"name",hash:{},data:t}):i))+'">\n'+(null!=(o=a.is_authenticated.call(s,{name:"is_authenticated",hash:{},fn:n.program(2,t,0),inverse:n.program(6,t,0),data:t}))?o:"")+'\n        <div class="survey-widgets survey-widgets-'+p((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,typeof i===c?i.call(s,{name:"name",hash:{},data:t}):i))+'">\n';return i=null!=(i=a.is_input||(null!=e?e.is_input:e))?i:u,r={name:"is_input",hash:{},fn:n.program(8,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a.is_input||(o=m.call(e,o,r)),null!=o&&(h+=o),h+="\n",i=null!=(i=a.is_textarea||(null!=e?e.is_textarea:e))?i:u,r={name:"is_textarea",hash:{},fn:n.program(15,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a.is_textarea||(o=m.call(e,o,r)),null!=o&&(h+=o),h+="\n",i=null!=(i=a.is_select||(null!=e?e.is_select:e))?i:u,r={name:"is_select",hash:{},fn:n.program(17,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a.is_select||(o=m.call(e,o,r)),null!=o&&(h+=o),h+"        </div>\n\n"+(null!=(o=a.if.call(s,null!=e?e.help_text:e,{name:"if",hash:{},fn:n.program(24,t,0),inverse:n.noop,data:t}))?o:"")+"      </div>\n\n"},2:function(n,e,a,l,t){var o;return null!=(o=a.is_submitter_name.call(null!=e?e:n.nullContext||{},{name:"is_submitter_name",hash:{},fn:n.noop,inverse:n.program(3,t,0),data:t}))?o:""},3:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=n.escapeExpression,p='          <label class="survey-label survey-label-'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'" for="survey-'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'">\n            '+c((i=null!=(i=a.prompt||(null!=e?e.prompt:e))?i:u,"function"==typeof i?i.call(s,{name:"prompt",hash:{},data:t}):i));return i=null!=(i=a.optional||(null!=e?e.optional:e))?i:u,r={name:"optional",hash:{},fn:n.program(4,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.optional||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(p+=o),p+"\n          </label>\n"},4:function(n,e,a,l,t){return" (optional)"},6:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=n.escapeExpression,p='        <label class="survey-label survey-label-'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'" for="survey-'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'">\n          '+c((i=null!=(i=a.prompt||(null!=e?e.prompt:e))?i:u,"function"==typeof i?i.call(s,{name:"prompt",hash:{},data:t}):i));return i=null!=(i=a.optional||(null!=e?e.optional:e))?i:u,r={name:"optional",hash:{},fn:n.program(4,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.optional||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(p+=o),p+"\n        </label>\n"},8:function(n,e,a,l,t){var o;return null!=(o=a.is_submitter_name.call(null!=e?e:n.nullContext||{},{name:"is_submitter_name",hash:{},fn:n.program(9,t,0),inverse:n.program(13,t,0),data:t}))?o:""},9:function(n,e,a,l,t){var o;return null!=(o=a.is_authenticated.call(null!=e?e:n.nullContext||{},{name:"is_authenticated",hash:{},fn:n.noop,inverse:n.program(10,t,0),data:t}))?o:""},10:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=n.escapeExpression,p='                <input id="survey-'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'" name="'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'" type="'+c((i=null!=(i=a.type||(null!=e?e.type:e))?i:u,"function"==typeof i?i.call(s,{name:"type",hash:{},data:t}):i))+'" ';return i=null!=(i=a.attrs||(null!=e?e.attrs:e))?i:u,r={name:"attrs",hash:{},fn:n.program(11,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.attrs||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(p+=o),p+' value="'+c(a.get_value.call(s,e,null!=e?e.name:e,{name:"get_value",hash:{},data:t}))+'">\n'},11:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return" "+s((o=null!=(o=a.key||(null!=e?e.key:e))?o:r,"function"==typeof o?o.call(i,{name:"key",hash:{},data:t}):o))+'="'+s((o=null!=(o=a.value||(null!=e?e.value:e))?o:r,"function"==typeof o?o.call(i,{name:"value",hash:{},data:t}):o))+'"'},13:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=n.escapeExpression,p='              <input id="survey-'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'" name="'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'" type="'+c((i=null!=(i=a.type||(null!=e?e.type:e))?i:u,"function"==typeof i?i.call(s,{name:"type",hash:{},data:t}):i))+'" ';return i=null!=(i=a.attrs||(null!=e?e.attrs:e))?i:u,r={name:"attrs",hash:{},fn:n.program(11,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.attrs||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(p+=o),p+' value="'+c(a.get_value.call(s,e,null!=e?e.name:e,{name:"get_value",hash:{},data:t}))+'">\n'},15:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=n.escapeExpression,p='          <textarea id="survey-'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'" name="'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'"';return i=null!=(i=a.attrs||(null!=e?e.attrs:e))?i:u,r={name:"attrs",hash:{},fn:n.program(11,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.attrs||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(p+=o),p+">"+c(a.get_value.call(s,e,null!=e?e.name:e,{name:"get_value",hash:{},data:t}))+"</textarea>\n"},17:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=n.escapeExpression,p='          <select id="survey-'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'" name="'+c((i=null!=(i=a.name||(null!=e?e.name:e))?i:u,"function"==typeof i?i.call(s,{name:"name",hash:{},data:t}):i))+'" type="'+c((i=null!=(i=a.type||(null!=e?e.type:e))?i:u,"function"==typeof i?i.call(s,{name:"type",hash:{},data:t}):i))+'" ';return i=null!=(i=a.attrs||(null!=e?e.attrs:e))?i:u,r={name:"attrs",hash:{},fn:n.program(11,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a.attrs||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(p+=o),p+">\n"+(null!=(o=a.select_item_value.call(s,e,null!=e?e.name:e,{name:"select_item_value",hash:{},fn:n.program(18,t,0),inverse:n.noop,data:t}))?o:"")+"          </select>\n"},18:function(n,e,a,l,t){var o,i,r,s="";return i=null!=(i=a.options||(null!=e?e.options:e))?i:a.helperMissing,r={name:"options",hash:{},fn:n.program(19,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(null!=e?e:n.nullContext||{},r):i,a.options||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(s+=o),s},19:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.label:e,{name:"if",hash:{},fn:n.program(20,t,0),inverse:n.program(22,t,0),data:t}))?o:""},20:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'            <option value="'+s((o=null!=(o=a.value||(null!=e?e.value:e))?o:r,"function"==typeof o?o.call(i,{name:"value",hash:{},data:t}):o))+'">'+s((o=null!=(o=a.label||(null!=e?e.label:e))?o:r,"function"==typeof o?o.call(i,{name:"label",hash:{},data:t}):o))+"</option>\n"},22:function(n,e,a,l,t){var o=n.lambda,i=n.escapeExpression;return'            <option value="'+i(o(e,e))+'">'+i(o(e,e))+"</option>\n"},24:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'        <div class="help-text place-'+s((o=null!=(o=a.name||(null!=e?e.name:e))?o:r,"function"==typeof o?o.call(i,{name:"name",hash:{},data:t}):o))+'-help-text">'+s((o=null!=(o=a.help_text||(null!=e?e.help_text:e))?o:r,"function"==typeof o?o.call(i,{name:"help_text",hash:{},data:t}):o))+"</div>\n"},26:function(n,e,a,l,t){var o=null!=e?e:n.nullContext||{},i=n.escapeExpression;return'      <p class="signed-in-as"><img src="'+i(a.current_user.call(o,"avatar_url",{name:"current_user",hash:{},data:t}))+'" class="avatar" /> '+i(a.current_user.call(o,"name",{name:"current_user",hash:{},data:t}))+'. <a class="" href="/users/logout/">Log Out</a></p>\n'},28:function(n,e,a,l,t){return""},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{},s=a.helperMissing,u=n.escapeExpression,c=n.lambda;return'<div id="reply">\n  <h4 class="reply-header">'+u((i=null!=(i=a.title||(null!=e?e.title:e))?i:s,"function"==typeof i?i.call(r,{name:"title",hash:{},data:t}):i))+'</h4>\n  <form id="new_response" class="new_response clearfix" action="#" method="post">\n    <input type="hidden" name="visible" value="on">\n    <input type="hidden" name="user_token" value="'+(null!=(i=null!=(i=a.user_token||(null!=e?e.user_token:e))?i:s,o="function"==typeof i?i.call(r,{name:"user_token",hash:{},data:t}):i)?o:"")+'">\n\n    <fieldset>\n'+(null!=(o=a.blockHelperMissing.call(e,c(null!=(o=null!=e?e.survey_config:e)?o.items:o,e),{name:"survey_config.items",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t}))?o:"")+'    </fieldset>\n\n    <input name="commit" type="submit" value="'+u(c(null!=(o=null!=e?e.survey_config:e)?o.submit_btn_text:o,e))+'" class="btn btn-primary submit-btn">\n\n'+(null!=(o=a.is_authenticated.call(r,{name:"is_authenticated",hash:{},fn:n.program(26,t,0),inverse:n.program(28,t,0),data:t}))?o:"")+'\n    <div class="form-spinner"></div>\n\n  </form>\n</div>\n'},useData:!0}),e["place-survey-response"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'  <hr>\n  <h5 style="text-align:center">Thanks for participating!</h5>\n'},useData:!0}),e.reporting=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return"<h1>What to look for</h1>"},useData:!0}),e["right-sidebar"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<div id="right-sidebar-spinner"></div>\n<a href="#" class="collapse-btn"></a>\n<div id="right-sidebar-content">\n\t<div id="sidebar-filter-menu"></div>\n  <ul id="sidebar-story-item-list"></ul>\n</div>'},useData:!0}),e["sidebar-story-detail"]=n({1:function(n,e,a,l,t){var o;return"\t\t"+n.escapeExpression((o=null!=(o=a.fullTitle||(null!=e?e.fullTitle:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"fullTitle",hash:{},data:t}):o))+"\n"},3:function(n,e,a,l,t){var o;return null!=(o=a.if.call(null!=e?e:n.nullContext||{},null!=e?e.title:e,{name:"if",hash:{},fn:n.program(4,t,0),inverse:n.program(6,t,0),data:t}))?o:""},4:function(n,e,a,l,t){var o;return"\t\t\t"+n.escapeExpression((o=null!=(o=a.title||(null!=e?e.title:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"title",hash:{},data:t}):o))+"\n"},6:function(n,e,a,l,t){var o;return"\t\t\t"+n.escapeExpression((o=null!=(o=a.name||(null!=e?e.name:e))?o:a.helperMissing,"function"==typeof o?o.call(null!=e?e:n.nullContext||{},{name:"name",hash:{},data:t}):o))+"\n"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{};return'<img class="sidebar-story-item-thumbnail" src='+n.escapeExpression((i=null!=(i=a.iconUrl||(null!=e?e.iconUrl:e))?i:a.helperMissing,"function"==typeof i?i.call(r,{name:"iconUrl",hash:{},data:t}):i))+' />\n<p class="sidebar-story-item-title">\n'+(null!=(o=a.if.call(r,null!=e?e.fullTitle:e,{name:"if",hash:{},fn:n.program(1,t,0),inverse:n.program(3,t,0),data:t}))?o:"")+"</p> "},useData:!0}),e.sidebar=n({1:function(n,e,a,l,t){var o,i=null!=e?e:n.nullContext||{},r=a.helperMissing,s=n.escapeExpression;return'        <li><a href="#'+s((o=null!=(o=a.id||(null!=e?e.id:e))?o:r,"function"==typeof o?o.call(i,{name:"id",hash:{},data:t}):o))+'-pane" role="tab"><i class="fa fa-'+s((o=null!=(o=a.icon||(null!=e?e.icon:e))?o:r,"function"==typeof o?o.call(i,{name:"icon",hash:{},data:t}):o))+'"></i></a></li>\n'},3:function(n,e,a,l,t){var o,i,r=null!=e?e:n.nullContext||{},s=a.helperMissing,u=n.escapeExpression;return'      <div class="sidebar-pane" id="'+u((i=null!=(i=a.id||(null!=e?e.id:e))?i:s,"function"==typeof i?i.call(r,{name:"id",hash:{},data:t}):i))+'-pane">\n        <h1 class="sidebar-header">'+u((i=null!=(i=a.title||(null!=e?e.title:e))?i:s,"function"==typeof i?i.call(r,{name:"title",hash:{},data:t}):i))+'<span class="sidebar-close"><i class="fa fa-chevron-left"></i></span></h1>\n        <div id="'+u((i=null!=(i=a.id||(null!=e?e.id:e))?i:s,"function"==typeof i?i.call(r,{name:"id",hash:{},data:t}):i))+'"></div>\n'+(null!=(o=a.if.call(r,null!=e?e.activity:e,{name:"if",hash:{},fn:n.program(4,t,0),inverse:n.noop,data:t}))?o:"")+"      </div>\n"},4:function(n,e,a,l,t){return'          <ul class="recent-points unstyled-list"></ul>\n'},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i=n.lambda,r=a.blockHelperMissing
+;return'<div id="sidebar" class="sidebar collapsed">\n  \x3c!-- TODO: Refactor these tabs/panes into a single view --\x3e\n  \x3c!-- Nav tabs --\x3e\n  <div class="sidebar-tabs">\n    <ul role="tablist">\n'+(null!=(o=r.call(e,i(null!=(o=null!=e?e.config:e)?o.panels:o,e),{name:"config.panels",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t}))?o:"")+'    </ul>\n  </div>\n\n  \x3c!-- Tab panes --\x3e\n  <div class="sidebar-content">\n\n'+(null!=(o=r.call(e,i(null!=(o=null!=e?e.config:e)?o.panels:o,e),{name:"config.panels",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t}))?o:"")+"\n  </div>\n\n</div\n"},useData:!0}),e.sponsors=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<h1>Sponsors</h1>\n\n<p>Seed funding provided by the Puget Sound Stewardship and Mitigation Fund, a community grantmaking fund created by Puget Soundkeeper Alliance and administered by the Rose Foundation for Communities and the Environment.</p>\n\n<p>Additional funding provided by Cascade Water Alliance in partnership with Sustainability Ambassadors.</p>\n\n<img src="/static/css/images/rose_logo.jpg" alt="Rose Foundation for Communities and the Environment - logo">\n\n<img src="/static/css/images/cascade_logo.jpg" alt="Cascade Water Alliance - logo">\n\n<img src="/static/css/images/sa-logo.jpg" alt="Sustainability Ambassadors - logo">\n\n<p class="btn btn-block btn-large close-btn">Back to the map</p>'},useData:!0}),e.stories=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<a href="/marra" rel="internal">\n<h1>Community Priorities</h1>\n<img src="http://i.imgur.com/eOhZTe6.png">\n</a>\n\n<a href="/cleanup" rel="internal">\n<h1>Pollution</h1>\n<img src="http://www.seattleglobalist.com/wp-content/uploads/2015/05/sign-700x467.jpg" width=100%>\n</a>\n\n<a href="/history" rel="internal">\n<h1>History</h1>\n<img src="http://duwamishcleanup.org/wp-content/uploads/2016/09/Duwamish1922-2.jpg">\n</a>'},useData:!0}),e.terms=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return"<h1>Terms and Conditions</h1>\n\n<p>By using this site you are agreeing to the terms and conditions presented herein. We are a grassroots community working to restore environmental health for present and future generations. Information that is deemed unproductive toward meeting this goal or hostile to the community may be removed at any time. We strive for accuracy but cannot guarantee the validity of information presented here.</p>"},useData:!0}),e.test=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return""},useData:!0}),e["unsupported-overlay"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<div class="unsupported-message">\n  <p class="browsehappy">You are using an unsupported browser. Please\n    <a href="http://browsehappy.com/">upgrade your browser</a> to improve\n    your experience.\n  </p>\n  <a href="#" class="close-btn btn-block close-unsupported-overlay">Proceed Anyway</a>\n</div>'},useData:!0}),e.why=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return"<h1>What is Shareabouts for?</h1>\n\n<p>Some ways you can use Shareabouts:</p>\n\n<ul>\n  <li>gather suggestions for possible locations for bike share stations</li>\n  <li>get comments and votes on proposed locations for food carts</li>\n  <li>collect data on vacant lots that could be urban farms</li>\n  <li>show infrastructure options and get public comment on each one</li>\n  <li>collect support for possible park locations</li>\n  <li>crowdsource issue reporting after a storm</li>\n</ul>\n"},useData:!0}),e.youthcorps2015=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<h1>Duwamish Valley Youth Corps 2015</h1>\n\n<p>We had a lot of fun and got a LOT of work done. Check out this highlight video!</p>\n\n<iframe src="https://player.vimeo.com/video/146331681?color=096e15&title=0&byline=0&portrait=0" width="500" height="282" frameborder="0" webkitallowfullscreen mozallowfullscreen allowfullscreen></iframe>'},useData:!0}),e["pages/about"]=n({1:function(n,e,a,l,t){return"Greensboro Participatory Budgeting:"},3:function(n,e,a,l,t){return"Our Community, Our Decision!"},5:function(n,e,a,l,t){return"Greensboro Participatory Budgeting is a democratic process in which community members from all five City Council districts are directly deciding how to spend $500,000 ($100,000 per district)."},7:function(n,e,a,l,t){return"From April to November 2017, volunteer budget delegates will work to develop these initial ideas into full project proposals. Greensboro residents will able to vote in October 2017 to determine the winning projects.\n"},9:function(n,e,a,l,t){return'For more information on participatory budgeting in Greensboro, visit <a href="http://www.greensboro-nc.gov/pb" target="_blank">greensboro-nc.gov/pb</a>.'},11:function(n,e,a,l,t){return'For information on the nationwide movement for PB, visit the website of the <a href="http://www.participatorybudgeting.org/" target="_blank">Participatory Budgeting Project'},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=a.blockHelperMissing,m="<h3>";return i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</h3>\n<h4>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</h4>\n<br>\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(5,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(7,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(9,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(11,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+'</a>.</p>\n\n<hr>\n<div class="clearfix" style="padding-top:0.5em;">\n  <a href="http://www.participatorybudgeting.org/" target="_blank"><img src="/static/pbp-logo.png" style="float:left;max-width:45%;"></a>\n</div>'},useData:!0}),e["pages/background"]=n({1:function(n,e,a,l,t){return"Background"},3:function(n,e,a,l,t){return"The Duwamish River is Seattle's only river, and is one of the most polluted waterways in the US. After 100 years of urban industrialization, it contains more than 40 toxic chemicals above safe limits for humans and the environment, including PCBs, dioxins, carcinogenic PAHs, and arsenic. Studies show that the low income communities living along the river have higher rates of heart disease, asthma, and other illnesses than other Seattle areas. It is also home to salmon, osprey, sea lions, and other creatures who face unsafe levels of contamination."},5:function(n,e,a,l,t){return"The history here is rich. The Duwamish Tribe lived along the shores of this river for over 10,000 years, since the last ice age, until they were removed by incoming settlers who came about a century and a half ago."},7:function(n,e,a,l,t){return"In 2001, the US Environmental Protection Agency declared the lower 5 miles of the Duwamish to be a superfund site, which mandates a cleanup paid by those responsible for the pollution. After more than a decade of public deliberation, the EPA came to a final decision in December 2014, and ordered a $342 million cleanup to take place over the next 17 years."},9:function(n,e,a,l,t){return"The responsible parties named so far are Boeing, King County, the City of Seattle, and the Port of Seattle, with several more expected to be named over the coming years."},11:function(n,e,a,l,t){return"We hope that 'Hey Duwamish' can be used as a community monitoring system to help ensure that the $342 million is well spent. We want to restore the health of our watershed to its maximum potential and highlight the progress being made through initiatives such as the <a href=\"http://murray.seattle.gov/murray-announces-funding-for-duwamish-river-community-projects\">Duwamish Opportunity Fund</a>."},13:function(n,e,a,l,t){return"Take me to the map!"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=a.blockHelperMissing,m="<h1>";return i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</h1>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(5,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(7,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(9,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(11,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+='</p>\n\n<p class="btn btn-block btn-large close-btn">',i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(13,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+"</p>\n"},useData:!0}),e["pages/build"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<h1>How to Build a Rain Garden</h1>\n<p><img class="aligncenter size-full wp-image-646" alt="diagram" src="http://www.12000raingardens.org/wp-content/uploads/2013/01/diagram.jpg" width="676" height="251" /></p>\n<p>If you&#8217;re interested in how to build a rain garden, a step-by-step process can be easily learned by attending a rain garden workshop, studying the <a title="WSU Rain Garden Handbook for Western Washington 2013" href="https://fortress.wa.gov/ecy/publications/publications/1310027.pdf" target="_blank">Washington State University Rain Garden Handbook for Homeowners</a> (18MB PDF download) and viewing the &#8220;How to build a rain garden WSU&#8221; online video (below). You will learn rain garden design and construction basics like:</p>\n<ul>\n<li>How to determine how much impervious surface you have, and how much of that water you want to manage.</li>\n<li>How to decide where on the property you&#8217;d like to build your rain garden.</li>\n<li>How to do a percolation test, to make sure the soils surrounding your rain garden plants can soak up the water.</li>\n<li>How to decide how big of a rain garden to create in the space you have available.</li>\n<li>How to implement rain garden construction next steps.</li>\n<li>How to know what the best rain garden plants are.</li>\n<li>How to maintain a rain garden.</li>\n</ul>\n<p>You don&#8217;t have to be a gardener or an engineer to build a rain garden, but it is important to consider your site carefully before you get started.</p>\n<p><iframe src="http://www.youtube.com/embed/9Kti4HJ45BM" height="315" width="560" allowfullscreen="" frameborder="0"></iframe></p>'},useData:!0}),e["pages/contribute"]=n({1:function(n,e,a,l,t){return"Get involved"},3:function(n,e,a,l,t){return"Want to help out? Awesome!"},5:function(n,e,a,l,t){return'The best way to get involved is to <a href="/new">submit a report</a> on the map, or comment on someone else\'s report.'},7:function(n,e,a,l,t){return'We could also use help writing code, designing a better interface, and promoting the project. <a href="http://github.com/smartercleanup/platform">Contribute to our repository on GitHub!</a>'},9:function(n,e,a,l,t){return"If you want to see a project like this in your community, let us know!"},11:function(n,e,a,l,t){return'<a href="http://eepurl.com/6IfA5">Send us a message</a> and we\'ll be in touch ;)'},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=a.blockHelperMissing,m="<h1>";return i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</h1>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(5,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(7,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(9,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(11,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+"\n"},useData:!0}),e["pages/features"]=n({1:function(n,e,a,l,t){return"Features"},3:function(n,e,a,l,t){return"For users,"},5:function(n,e,a,l,t){return"a simple map interface, to explore submitted locations and add your own"},7:function(n,e,a,l,t){return"site design that works well on small screens"},9:function(n,e,a,l,t){return"For designers and developers,"},11:function(n,e,a,l,t){return"flexible text and layout, making it easy to translate the user interface text or customize the sentences used"},13:function(n,e,a,l,t){return"geospatial database, so you can limit contributions to within boundaries, or asiggn neighborhood names"},15:function(n,e,a,l,t){return"For project managers during a project,"},17:function(n,e,a,l,t){return"simple admin interface"},19:function(n,e,a,l,t){return"blogging tools to create pages or blog posts"},21:function(n,e,a,l,t){return"map and location configuration"},23:function(n,e,a,l,t){return"export to GIS"},25:function(n,e,a,l,t){return"simple status dashboard"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=a.blockHelperMissing,m="<h1>";return i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</h1>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<ul>\n  <li>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(5,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="\n  <li>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(7,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="\n</ul>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(9,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<ul>\n  <li>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(11,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="\n  <li>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(13,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="\n</ul>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(15,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<ul>\n  <li>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(17,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="\n  <li>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(19,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="\n  <li>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(21,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="\n  <li>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(23,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="\n  <li>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(25,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+"\n</ul>\n"},useData:!0}),e["pages/getinvolved"]=n({1:function(n,e,a,l,t){return"Get more involved"},3:function(n,e,a,l,t){return"Already submitted an idea? There are many more ways to get involved in participatory budgeting:"},5:function(n,e,a,l,t){return"Volunteer"},7:function(n,e,a,l,t){return'You can help to get the word out to district residents, facilitate small group meetings, provide support at PB events, and more. Sign up <a href="http://www.123contactform.com/form-2317799/PB-Oakland-Volunteer-Sign-Up" target="_blank">here</a>.'},9:function(n,e,a,l,t){return"Contact Us"},11:function(n,e,a,l,t){return"Have a question? Feel free to contact us."},13:function(n,e,a,l,t){return'<iframe src="http://www.123contactform.com/form-1169694/PB-Oakland-Contact-Us" height="400" width=100% </iframe>'},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=a.blockHelperMissing,m="<h1>";return i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</h1>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<h3>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(5,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</h3>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(7,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<h3>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(9,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</h3>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(11,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(13,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m},useData:!0}),e["pages/greenscreen"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'    <p>Hello residents of Georgetown and South Park,</p>\n    \n    <p>Just Health Action (JHA), a non-profit Seattle-based organization\n    received a grant from King County Waste Water Treatment Division to build\n    an actionable project in the neighborhood to mitigate air pollution.  This\n    action-oriented project is called Green Solutions to Air Pollution and will\n    leverage the existing resources and work achieved through the Duwamish\n    Community Action for Clean Air project coordinated by Duwamish River\n    Cleanup Coalition (DRCC). We will be partnering with the DRCC and Urban\n    Systems Design’s program “Duwamish Infrastructure Restoration Training”\n    (DIRT Corps) to implement the project.\n     \n<p>After talking to a few residents at community workshops this fall about different types of actionable projects that could reduce air pollution in the neighborhoods, we heard you say that installing a “green screen” was the most exciting option (a detailed description of what a green screen is\nbelow). We now need your feedback about where you would like this green screen to go.  Though we have spoken to some residents already, we want to make sure we hear as many voices as possible before deciding on the final locations (one location per neighborhood). More information about the details of the project are listed below the instructions.</p>\n    \n    <p><strong>To add your green screen location nomination to the map:</strong></p>\n\n<ul>\n<li>Hit the orange button that says \'Add a report to the map\'</li>\n<li>Move the red pin to mark your exact desired location (by dragging and zooming the map)</li>\n<li>Select <em>Green Screen Vote</em> under the dropdown menu where it says \'type of report.\'</li>\n</ul>\n\n    <a href="#" id="add-place" class="btn btn-block btn-primary btn-large"><span>Add a report to the map!</span></a>\n    <img src="/static/css/images/marker-drop_screenshot.jpg" alt="Location pin">\n    <img src="/static/css/images/screenshot_greenscreen_dropdown.jpg" alt="Green Screen Vote - Dropdown Menu">\n  <br>    \n    <p>Feel free to comment on other Green Screen votes that others have chosen\n    to help guide the community process. To comment on other nominations just\n    click on the pin and a comment window will open to the right.</p>\n    \n    For those folks not familiar with the project, here is some more\n    information.</p>\n<img src="/static/css/images/greenscreen_example.png" alt="Green screen example" style="float:left;margin:0 5px 0 0;" />  \n    <h3>What is a green screen?</h3>\n    <p>A green screen is a metal frame structure that will support the growth\n    of native and fast growing vines (no invasive species will be used).  In\n    two to three years the entire structure will be completely covered by the\n    vines, and as the vines grow, so will their ability to absorb air\n    pollution.  It will end up looking like a ten-foot tall and twenty to\n    thirty foot long wall of green growing things.  An example of a green\n    screen on the Goat Hill Parking Garage in Seattle can be found <a href="http://greenscreen.com/docs/Projects%20PDFs/All%20Individual%20Project%20PDFs/Goat%20Hill%20Parking.pdf">here</a>.\n\n    <h3>Why a “green screen”?</h3>\n    <p>A green screen is a great tool to help absorb air pollution.  It is a\n    version of the green wall or living wall concept that requires very little\n    maintenance, so it has the potential to be sustainable for many years to\n    come.  The project is something that can be constructed relatively quickly\n    on a narrow piece of land that cannot accommodate trees. The green screen\n    will look attractive once it’s grown in, as well as reduce noise pollution,\n    absorb water, reduce heat, create local jobs, and it is generally good for\n    your health. Here is a link to more information about the benefits of\n    living walls general:\n    <a href="http://www.greenovergrey.com/green-wall-benefits/overview.php">http://www.greenovergrey.com/green-wall-benefits/overview.php</a></p>\n     \n    <h3>What are we asking you to do?</h3>\n    <p>We would like your input on where you think this green screen should be\n    placed.  To add your voice, go to HeyDuwamish.org and place a pin on\n    the map where you would like the green screen to go (see above for more\n    detailed instructions).  You can also add a comment along with your map pin\n    on why you picked this spot, and/or comment on other people’s pins. \n     Please comment by the end of December. </p>\n     \n    <h3>What is the timeline and how will the project get built?</h3>\n    <p>Two projects will be built, one in Georgetown, and one in South Park. \n    During the first week of January, we will take all of the pins and\n    suggestions into account, and decide on a location for the green screen. \n    We will then negotiate with the landowner or the city as the case may be to\n    site the project on their property. We are aiming to begin construction in\n    the winter or spring of 2016.  The project will be built by the Duwamish\n    Valley Youth Corps, and the Duwamish Infrastructure Restoration Training\n    Corps (DIRT Corps)</p>\n     \n    <p>If you have any additional questions, or would like more information,\n    you can contact us at:</p>\n    <p>Linn Gould - gouldjha@gmail.com</p>\n    <p>Andrew Schiffer - bricktree66@gmail.com</p>\n\n<img src="/static/css/images/kclogo.gif" alt="King County">\n<img src="/static/css/images/jha-logo.jpg" alt="Just Health Action">\n<img src="/static/css/images/dirtcorps-logo.jpg" alt="Dirt Corps">\n<img src="/static/css/images/drcc-logo_small.jpg" alt="Just Health Action">'},useData:!0}),e["pages/guidelines"]=n({1:function(n,e,a,l,t){return'<iframe src="https://player.vimeo.com/video/162743651?color=ff9933&byline=0" width="640" height="360" frameborder="0" webkitallowfullscreen mozallowfullscreen allowfullscreen></iframe>\n'},3:function(n,e,a,l,t){return'<h2>How does PB work in Oakland?</h2>\n<p>\nParticipatory Budgeting (PB) is an innovative democratic process that gives Oakland residents of City Council Districts 1 and 2 the power to set priorities for how federal Community Development Block Grant (CDBG) funds should be spent to improve low-to-moderate income communities in their districts. The projects that receive the most votes up to the maximum dollar amount will become the CDBG priorities for the district. In December and January, residents came together at events in their districts and online to discuss community needs and brainstorm ideas for funded projects that would benefit low-to-moderate incomes communities. In early February, volunteer budget delegates evaluated and prioritized those ideas and turned them into the priority projects that will appear on the PB ballot from February 18-22, 2017.\n</p>\n<p>\nFollowing the vote, the City of Oakland will release a Request for Proposals (RFP), and city and non-profit agencies may apply for funds to implement priority projects. Community Development District Board Members in Districts 1 and 2 will review all applications in response to the RFP and make recommendations to the city on which should be funded. For more information on the RFP process, <a href="http://www2.oaklandnet.com/government/o/hcd/s/CDBG/index.htm">visit the City of Oakland\'s CDBG website</a>.\n</p>\n<p>\n<strong>In Council District 1</strong>, the 2017-19 CDBG allocation is $113,654 per year over two years, or $227,308 in total. For a map of the CDBG eligible area, <a href="http://www2.oaklandnet.com/oakca1/groups/ceda/documents/webcontent/dowd009136.pdf">click here</a>.\n</p>\n<p>\n<strong>In Council District 2</strong>, the 2017-19 CDBG allocation is $278,685 per year over two years, or $557,370 in total. For a map of the CDBG eligible area, <a href="http://www2.oaklandnet.com/oakca1/groups/ceda/documents/webcontent/dowd009137.pdf">click here</a>.\n</p>\n<p>\nNot sure if you live in District 1 or 2? Enter your address on <a href="https://participatorybudgeting.nationbuilder.com/r?u=http%3A%2F%2Fmapgis.oaklandnet.com%2Fcouncildistricts%2F&e=1860e7904dcd651a71dfc072d7cae68cb859e106&utm_source=participatorybudgeting&utm_medium=email&utm_campaign=oakd1vote&n=7&test_email=1">this map</a> to find out.\n</p>\n<em><strong>\nFor a full description of the PB process in Oakland, download the <a href="https://drive.google.com/file/d/0B8FaQ6MQKJm1U0xQSVNUQ1ZEZk0/view?usp=sharing">PB Oakland 2016-17 Rulebook</a>.\n</em>\n</strong>\n'},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=a.blockHelperMissing,p="";return i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a._||(o=c.call(e,o,r)),null!=o&&(p+=o),i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a._||(o=c.call(e,o,r)),null!=o&&(p+=o),p},useData:!0}),e["pages/instructions"]=n({1:function(n,e,a,l,t){return"Detailed instructions"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s="<h1>";return i=null!=(i=a._||(null!=e?e._:e))?i:a.helperMissing,r={name:"_",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(null!=e?e:n.nullContext||{},r):i,a._||(o=a.blockHelperMissing.call(e,o,r)),null!=o&&(s+=o),s+"</h1>"},useData:!0}),e["pages/involved"]=n({1:function(n,e,a,l,t){return"Get More Involved"},3:function(n,e,a,l,t){return"Already submitted an idea? There are many more ways to get involved in participatory budgeting:"},5:function(n,e,a,l,t){return"Attend a Neighborhood Assembly"},7:function(n,e,a,l,t){return" Each district will have a neighborhood assembly and other, informal idea collection events, to encourage widespread participation."},9:function(n,e,a,l,t){return"Click here</a> for upcoming assembly dates, times, and locations."},11:function(n,e,a,l,t){return"Connect with Us"},13:function(n,e,a,l,t){return'Find us on <a href="https://www.facebook.com/PBGreensboro/" target="_blank"> Facebook</a> and follow us on <a href="https://twitter.com/GreensboroPB" target="_blank"> Twitter'},15:function(n,e,a,l,t){return"Contact Us"},17:function(n,e,a,l,t){return'<strong>Have a question?</strong><br>\nPlease contact Valerie Warren, Community Outreach Coordinator for Participatory Budgeting Greensboro, by email at <a href="mailto:valerie@participatorybudgeting.org">valerie@participatorybudgeting.org</a> or phone at 336‑373‑7750.\n'},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=a.blockHelperMissing,m="<h1>";return i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</h1>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<h5>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(5,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</h5>\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(7,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+=' <a href="http://www.greensboro-nc.gov/index.aspx?page=5084">',i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(9,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n\n<h5>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(11,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</h5>\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(13,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</a>!</p>\n\n<h5>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(15,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</h5>\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(17,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),
+null!=o&&(m+=o),m+"</p>"},useData:!0}),e["pages/overview"]=n({1:function(n,e,a,l,t){return"Welcome"},3:function(n,e,a,l,t){return"We are here to educate and inspire the environmental stewardship of Longfellow Creek in West Seattle, starting with Roxhill Bog, all the way down to Elliot Bay!"},5:function(n,e,a,l,t){return"Start Mapping!"},7:function(n,e,a,l,t){return'This project made possible by our generous <a href="/page/sponsors">sponsors'},9:function(n,e,a,l,t){return"We are here to educate and inspire action for our Central Puget Sound watersheds"},11:function(n,e,a,l,t){return"Cedar River / Lake Washington Watershed"},13:function(n,e,a,l,t){return"Green / Duwamish Watershed"},15:function(n,e,a,l,t){return"Puyallup / White Watershed"},17:function(n,e,a,l,t){return'This project made possible by our generous <a href="/page/sponsors" rel="internal">sponsors'},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c="function",p=a.blockHelperMissing,m="<h1>";return i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</h1>\n\n\x3c!-- <p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+='</p>\n<a href="/longfellow" rel="internal">\n<img src="https://5937delridge.files.wordpress.com/2011/10/beaver-pond1.jpg">\n<p class="btn btn-block btn-large close-btn">',i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(5,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</p>\n</a>\n<br>\n<br>\n",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(7,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+="</a>. --\x3e\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(9,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+='</p>\n\x3c!-- \n<figure class="snip1579">\n\t<img src="/static/css/images/overview-image-1.jpg"/>\n\t<figcaption>\n\t  <h3>Lake Sammamish Watershed</h3>\n\t  <blockquote>\n\t    <p></p>\n\t  </blockquote>\n\t</figcaption><a href="/sammamish" rel="internal"></a>\n</figure>\n<figure class="snip1579">\n\t<img src="/static/css/images/overview-image-2.jpg"/>\n  <figcaption>\n    <h3>',i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(11,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+='</h3>\n    <blockquote>\n      <p></p>\n    </blockquote>\n  </figcaption><a href="/cedar" rel="internal"></a>\n</figure>\n<figure class="snip1579">\n\t<img src="/static/css/images/overview-image-3.jpg" alt="tools"/>\n  <figcaption>\n    <h3>',i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(13,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+='</h3>\n    <blockquote>\n      <p></p>\n    </blockquote>\n  </figcaption><a href="/green-duwamish" rel="internal"></a>\n</figure>\n<figure class="snip1579">\n  <img src="/static/css/images/overview-image-4.jpg" alt="tools"/>\n  <figcaption>\n    <h3>',i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(15,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+='</h3>\n    <blockquote>\n      <p></p>\n    </blockquote>\n  </figcaption><a href="/puyallup-white" rel="internal"></a>\n</figure>\n<br><br>\n --\x3e\n\n<figure class="flow-layout flow-sq flow-sq-left">\n  <a href="/cedar" rel="internal">\n    <img src="/static/css/images/watershed-cedar-sq.jpg"/>\n    <figcaption class="flow-layout-title flow-layout-title-bottom">\n      <h3>Cedar River / Lake Washington Watershed</h3>\n    </figcaption>\n  </a>\n</figure>\n\n<figure class="flow-layout flow-vert flow-vert-right">\n  <a href="/green-duwamish" rel="internal">\n    <img src="/static/css/images/watershed-green-duwamish-vert-b.jpg"/>\n    <figcaption class="flow-layout-title flow-layout-title-bottom">\n      <h3>Green / Duwamish Watershed</h3>\n    </figcaption>\n  </a>\n</figure>\n\n<figure class="flow-layout flow-sq flow-sq-left">\n  <a href="/sammamish" rel="internal">\n    <img src="/static/css/images/watershed-sammamish-alt-sq.jpg"/>\n    <figcaption class="flow-layout-title flow-layout-title-bottom">\n      <h3>Lake Sammamish Watershed</h3>\n    </figcaption>\n  </a>\n</figure>\n\n<div class="clearfix"></div>\n\n<figure class="flow-layout">\n  <a href="/puyallup-white" rel="internal">\n    <img src="/static/css/images/watershed-puyallup-alt-horiz.png"/>\n    <figcaption class="flow-layout-title flow-layout-title-bottom">\n      <h3>Puyallup / White Watershed</h3>\n    </figcaption>\n  </a>\n</figure>\n\n<br />\n\n',i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(17,t,0),inverse:n.noop,data:t},o=typeof i===c?i.call(s,r):i,a._||(o=p.call(e,o,r)),null!=o&&(m+=o),m+"</a>."},useData:!0}),e["pages/reporting"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return"<h1>What to look for</h1>"},useData:!0}),e["pages/sponsors"]=n({1:function(n,e,a,l,t){return"Sponsors"},3:function(n,e,a,l,t){return"Seed funding provided by the Puget Sound Stewardship and Mitigation Fund, a community grantmaking fund created by Puget Soundkeeper Alliance and administered by the Rose Foundation for Communities and the Environment."},5:function(n,e,a,l,t){return"Back to the map"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=a.blockHelperMissing,p="<h1>";return i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a._||(o=c.call(e,o,r)),null!=o&&(p+=o),p+="</h1>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a._||(o=c.call(e,o,r)),null!=o&&(p+=o),p+='</p>\n\n<p>Additional funding provided by Cascade Water Alliance in partnership with Sustainability Ambassadors.</p>\n\n<img src="/static/css/images/rose_logo.jpg" alt="Rose Foundation for Communities and the Environment - logo">\n\n<img src="/static/css/images/cascade_logo.jpg" alt="Cascade Water Alliance - logo">\n\n<img src="/static/css/images/sa-logo.jpg" alt="Sustainability Ambassadors - logo">\n\n<p class="btn btn-block btn-large close-btn">',i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(5,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a._||(o=c.call(e,o,r)),null!=o&&(p+=o),p+"</p>"},useData:!0}),e["pages/stories"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return'<a href="/marra" rel="internal">\n<h1>Community Priorities</h1>\n<img src="http://i.imgur.com/eOhZTe6.png">\n</a>\n\n<a href="/cleanup" rel="internal">\n<h1>Pollution</h1>\n<img src="http://www.seattleglobalist.com/wp-content/uploads/2015/05/sign-700x467.jpg" width=100%>\n</a>\n\n<a href="/history" rel="internal">\n<h1>History</h1>\n<img src="http://duwamishcleanup.org/wp-content/uploads/2016/09/Duwamish1922-2.jpg">\n</a>'},useData:!0}),e["pages/terms"]=n({1:function(n,e,a,l,t){return"Terms and Conditions"},3:function(n,e,a,l,t){return"By using this site you are agreeing to the terms and conditions presented herein. We are a grassroots community working to restore environmental health for present and future generations. Information that is deemed unproductive toward meeting this goal or hostile to the community may be removed at any time. We strive for accuracy but cannot guarantee the validity of information presented here."},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=a.blockHelperMissing,p="<h1>";return i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a._||(o=c.call(e,o,r)),null!=o&&(p+=o),p+="</h1>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a._||(o=c.call(e,o,r)),null!=o&&(p+=o),p+"</p>"},useData:!0}),e["pages/test"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return""},useData:!0}),e["pages/why"]=n({compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){return"<h1>What is Shareabouts for?</h1>\n\n<p>Some ways you can use Shareabouts:</p>\n\n<ul>\n  <li>gather suggestions for possible locations for bike share stations</li>\n  <li>get comments and votes on proposed locations for food carts</li>\n  <li>collect data on vacant lots that could be urban farms</li>\n  <li>show infrastructure options and get public comment on each one</li>\n  <li>collect support for possible park locations</li>\n  <li>crowdsource issue reporting after a storm</li>\n</ul>\n"},useData:!0}),e["pages/youthcorps2015"]=n({1:function(n,e,a,l,t){return"Duwamish Valley Youth Corps 2015"},3:function(n,e,a,l,t){return"We had a lot of fun and got a LOT of work done. Check out this highlight video!"},compiler:[7,">= 4.0.0"],main:function(n,e,a,l,t){var o,i,r,s=null!=e?e:n.nullContext||{},u=a.helperMissing,c=a.blockHelperMissing,p="<h1>";return i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(1,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a._||(o=c.call(e,o,r)),null!=o&&(p+=o),p+="</h1>\n\n<p>",i=null!=(i=a._||(null!=e?e._:e))?i:u,r={name:"_",hash:{},fn:n.program(3,t,0),inverse:n.noop,data:t},o="function"==typeof i?i.call(s,r):i,a._||(o=c.call(e,o,r)),null!=o&&(p+=o),p+'</p>\n\n<iframe src="https://player.vimeo.com/video/146331681?color=096e15&title=0&byline=0&portrait=0" width="500" height="282" frameborder="0" webkitallowfullscreen mozallowfullscreen allowfullscreen></iframe>'},useData:!0})}();
